replace feedback_pipeline.py and remove old files
