--- base-example.yaml
@@ -1,16 +1,0 @@
----
-document: feeback-pipeline-base
-version: 1
-data:
-  id: fedora-container-base
-  name: Fedora Container Base
-  versions:
-    '30':
-      packages:
-      - package1
-      - package2
-      options:
-      - no-docs
-      - no-weak-deps
-      source:
-        releasever: 30
