+++ feedback_pipeline.py
@@ -1,1098 +1,64 @@
 #!/usr/bin/python3
 
-import argparse, yaml, tempfile, os, subprocess, json, jinja2, datetime, copy, re, dnf
+import argparse, yaml, tempfile, os, subprocess, json, jinja2, datetime, copy, re, dnf, pprint
 import rpm_showme as showme
 
 
-# Data types:
+# Features of this new release
+# - multiarch from the ground up!
+# - more resilient
+# - better internal data structure
+# - user-defined views
+
+
+###############################################################################
+### Help ######################################################################
+###############################################################################
+
+
+# Configs:
+#   TYPE:           KEY:          ID:
+# - repo            repos         repo_id
+# - env_conf        envs          env_id
+# - workload_conf   workloads     workload_id
+# - label           labels        label_id
+# - conf_view       views         view_id
 #
-#   Base = {
-#       "id": "fedora-container-base",
-#       "name": "Fedora Container Base",
-#       "versions": {
-#           "30": {
-#               "packages": [
-#                   "package-name",
-#                   "package-name"
-#               ],
-#               "options": [
-#                   "no-docs",
-#                   "no-weak-deps"
-#               ],
-#               "source": {
-#                   "releasever": 30
-#               }
-#           },
-#           "31": {...}
-#       }
-#   }
+# Data:
+#   TYPE:         KEY:                 ID:
+# - pkg           pkgs/repo_id/arch    NEVR
+# - env           envs                 env_id:repo_id:arch_id
+# - workload      workloads            workload_id:env_id:repo_id:arch_id
+# - view          views                view_id:repo_id:arch_id
 #
-#   UseCase = {
-#       "id": "unique-id",
-#       "name": "Human-friendly name",
-#       "packages": [
-#           "package-name",
-#           "package-name"
-#       ],
-#       "options": [
-#           "no-docs",
-#           "no-weak-deps"
-#       ],
-#       "install_on": [
-#           "fedora-container-base:30",
-#           "fedora-container-base:31",
-#           "empty:30",
-#           "empty:31"
-#       ]
-#   }
 #
-#   BaseInstallation = {
-#       "id": "fedora-container-base:30",
-#       "base_id": "fedora-container-base",
-#       "base_version": "30",
-#       "packages": [
-#           Package,    # Package from rpm-showme
-#           Package     # Package from rpm-showme
-#       ]
-#   }
 #
-#   UseCaseInstallation = {
-#       "id": "unique-id:fedora-container-base:30",
-#       "use_case_id: "unique-id",
-#       "base_id": "fedora-container-base",
-#       "base_version": "30,
-#       "packages": [
-#           Package,    # Package from rpm-showme
-#           Package     # Package from rpm-showme
-#       ]
-#   }
+
+
+
+###############################################################################
+### Some initial stuff ########################################################
+###############################################################################
+
+# Error in global settings for Feedback Pipeline
+# Settings to be implemented, now hardcoded below
+class SettingsError(Exception):
+    pass
+
+# Error in user-provided configs
+class ConfigError(Exception):
+    pass
+
+# Error in downloading repodata
+class RepoDownloadError(Exception):
+    pass
+
 
 def log(msg):
     print(msg)
 
 def err_log(msg):
     print("ERROR LOG:  {}".format(msg))
-
-def get_configs(directory):
-    configs = {}
-    configs["bases"] = {}
-    configs["use_cases"] = {}
-
-    for yml_file in os.listdir(directory):
-        if not yml_file.endswith(".yaml"):
-            continue
-
-        with open(os.path.join(directory, yml_file), "r") as file:
-            try:
-                document = yaml.safe_load(file)
-            except yaml.YAMLError as err:
-                print(err)
-
-            if not ("document" in document and "version" in document):
-                print("Error: {file} is invalid.".format(file=yml_file))
-
-            if document["document"] == "feeback-pipeline-base":
-    
-                base = {}
-                base["id"] = document["data"]["id"]
-                base["name"] = document["data"]["name"]
-                base["versions"] = {}
-                for version,version_data in document["data"]["versions"].items():
-                    base["versions"][version] = {}
-
-                    packages = []
-                    if "packages" in version_data:
-                        packages = version_data["packages"]
-                    base["versions"][version]["packages"] = packages
-
-                    options = []
-                    if "options" in version_data:
-                        options = version_data["options"]
-                    base["versions"][version]["options"] = options
-
-                    releasever = version_data["source"]["releasever"]
-                    base["versions"][version]["source"] = {}
-                    base["versions"][version]["source"]["releasever"] = str(releasever)
-
-                configs["bases"][base["id"]] = base
-
-            if document["document"] == "feeback-pipeline-use-case":
-
-                use_case = {}
-                use_case["id"] = document["data"]["id"]
-                use_case["name"] = document["data"]["name"]
-
-                packages = []
-                if "packages" in document["data"]:
-                    packages = document["data"]["packages"]
-                use_case["packages"] = packages
-
-                options = []
-                if "options" in document["data"]:
-                    options = document["data"]["options"]
-                use_case["options"] = options
-
-                use_case["install_on"] = document["data"]["install_on"]
-
-                configs["use_cases"][use_case["id"]] = use_case
-
-    return configs
-
-def _install_packages(installroot, pkgs_to_install, options, releasever, cachedir=None, installing_base_image=False, empty_base=False):
-
-    # Explanation of installing_base_image:
-    #  - There are two types of installation:
-    #       1/ Base image
-    #       2/ Use case
-    #  - Base image installation needs to start fresh in an empty installrot,
-    #    and the result needs to be written into RPMDB, so a use case installation
-    #    can then consume it. It can't read it, though, because it doesn't exist.
-    #  - Use case installation needs to then read it, but it won't write it again
-    #    because for writing it, it would need to download all packages. A high
-    #    and unneccessary price to pay. 
-
-    # Some logging
-    log("    installroot:     {}".format(installroot))
-    log("    cachedir:        {}".format(cachedir))
-    log("    number of pkgs:  {}".format(len(pkgs_to_install)))
-    log("    empty base:      {}".format(empty_base))
-
-    with dnf.Base() as base:
-
-        if cachedir:
-            base.conf.cachedir = cachedir
-
-        base.conf.substitutions['releasever'] = releasever
-        if "no-weak-deps" in options:
-            base.conf.install_weak_deps = False
-
-        base.conf.tsflags = []
-        if "no-docs" in options:
-            base.conf.tsflags.append('nodocs')
-        # Base image? We're writing things down. This makes the transaction to only write
-        # them down without installing to save time and resources.
-        if installing_base_image:
-            base.conf.tsflags.append('justdb')
-
-        base.conf.installroot = installroot
-
-        log("    Reading repos...")
-        base.read_all_repos()
-
-        # Base image? Starting fresh == don't read the local RPMBD.
-        # Use case? Load it. But only if the base is not empty, otherwise there would be an OSError
-        # about not being able to read the RPMDB (it doesn't get created for empty base images).
-        if installing_base_image or empty_base:
-            base.fill_sack(load_system_repo=False)
-        else:
-            base.fill_sack(load_system_repo=True)
-
-        log("    Adding packages to the install list...")
-        for pkg in pkgs_to_install:
-            # the following can throw dnf.exceptions.MarkingError
-            # when the requested package doesn't exist
-            try:
-                base.install(pkg)
-            except dnf.exceptions.MarkingError:
-                err_log("Package '{}' could not be found!".format(pkg))
-                continue
-
-        # the following can throw dnf.exceptions.DepsolveError
-        # if dependencies can't be resolved     
-        log("    Resolving dependnecies...")   
-        base.resolve()
-
-        if installing_base_image:
-            log("    Base image dependency resolution.")
-            # Write the result intoi RPMDB.
-            # The transaction needs us to download all the packages. :(
-            # So let's do that to make it happy.
-            log("    Downloading packages...")
-            base.download_packages(base.transaction.install_set)
-            log("    Running transaction...")
-            base.do_transaction()
-            # The query in this case is just the resolve install set. Easy peasy!
-            log("    Creating the query object...")
-            query = base.sack.query().filterm(pkg=base.transaction.install_set)
-        
-        else:
-            log("    Use case dependency resolution.")
-            # The query here is the combinaton of the pre-installed base image
-            # and the resolved install set. So let's get both of those,
-            # merge them, and make a query out of that!
-            log("    Creating the query object...")
-            query_base = base.sack.query()
-            base_installed = set(query_base.installed())
-            use_case_installed = set(base.transaction.install_set)
-            all_installed = set.union(base_installed, use_case_installed)
-            query = base.sack.query().filterm(pkg=all_installed)
-
-        # And finally, the reason we went through all that trouble, was to get
-        # some interesting data! So let's finally get it.
-        log("    Saving package data...")
-        packages = {}
-        for pkg in query:
-            package = {}
-            package["name"] = pkg.name
-            package["epoch"] = pkg.epoch
-            package["version"] = pkg.version
-            package["release"] = pkg.release
-            package["arch"] = pkg.arch
-            package["nevra"] = str(pkg)
-            package["size"] = pkg.installsize
-            package["requires"] = []
-            package["requires_resolved"] = []
-            package["recommends"] = []
-            package["recommends_resolved"] = []
-            package["suggests"] = []
-            package["suggests_resolved"] = []
-
-            for req in pkg.requires:
-                package["requires"].append(str(req))
-
-            for req in pkg.recommends:
-                package["recommends"].append(str(req))
-
-            for req in pkg.suggests:
-                package["suggests"].append(str(req))
-
-            deps = query.filter(provides=pkg.requires)
-            for dep in deps:
-                package["requires_resolved"].append(dep.name)
-
-            deps = query.filter(provides=pkg.recommends)
-            for dep in deps:
-                package["recommends_resolved"].append(dep.name)
-
-            deps = query.filter(provides=pkg.suggests)
-            for dep in deps:
-                package["suggests_resolved"].append(dep.name)
-
-            packages[package["name"]] = package
-
-        log("    DONE!")
-        log("")
-    return packages
-
-
-def install_and_load(configs):
-    timestamp = str(datetime.datetime.now().strftime("%d/%m/%Y %H:%M"))
-
-    installs = {}
-    installs["bases"] = {}
-    installs["use_cases"] = {}
-    installs["timestamp"] = timestamp
-
-    bases = configs["bases"]
-    use_cases = configs["use_cases"]
-
-    with tempfile.TemporaryDirectory() as root:
-        # Step 1: Let's do all the base images!
-        for base_id, base in bases.items():
-            for base_version, base_version_data in base["versions"].items():
-                base_install_id = "{id}:{version}".format(
-                        id=base_id, version=base_version)
-
-                # Where to install
-                # We're writing data in the installroot, so each base image
-                # needs its own space.
-                dirname = "{id}--{version}".format(
-                        id=base_id, version=base_version)
-                installroot = os.path.join(root, dirname)
-
-                # What and how to install
-                packages = base_version_data["packages"]
-                options = base_version_data["options"]
-                releasever = base_version_data["source"]["releasever"]
-
-                # DNF cache directory
-                # Loading repodata only once speeds things up quite a bit
-                cachedir = os.path.join(root, "dnf-cache-{releasever}".format(releasever=releasever))
-
-                # Log!
-                log("Analyzing a base image: {name} {release} ({number_pkgs} packages)".format(
-                        name=base["name"],
-                        release=base_version,
-                        number_pkgs=len(packages)
-                ))
-
-                # Do the installation
-                installed_packages = _install_packages(installroot=installroot,
-                                                       pkgs_to_install=packages,
-                                                       options=options,
-                                                       releasever=releasever,
-                                                       cachedir=cachedir,
-                                                       installing_base_image=True)
-
-                # Save
-                base_installation = {}
-                base_installation["id"] = base_install_id
-                base_installation["base_id"] = base_id
-                base_installation["base_version"] = base_version
-                base_installation["packages"] = installed_packages
-                installs["bases"][base_install_id] = base_installation
-
-        # Step 2: Let's do all the use cases on top of the base images!
-        for use_case_id, use_case in use_cases.items():
-            for base_install_id in use_case["install_on"]:
-                use_case_install_id = "{use_case_id}:{base_install_id}".format(
-                        use_case_id=use_case_id,
-                        base_install_id=base_install_id)
-
-                # Where to install
-                # Since we're not writing anything into the installroot,
-                # let's just use the base image's installroot!
-                base_id = installs["bases"][base_install_id]["base_id"]
-                base_version = installs["bases"][base_install_id]["base_version"]
-                base_dirname = "{base_id}--{base_version}".format(
-                        base_id=base_id,
-                        base_version=base_version)
-                base_installroot = os.path.join(root, base_dirname)
-
-                # What and how to install
-                packages = use_case["packages"]
-                options = use_case["options"]
-                releasever = bases[base_id]["versions"][base_version]["source"]["releasever"]
-
-                # DNF cache directory
-                # Loading repodata only once speeds things up quite a bit
-                cachedir = os.path.join(root, "dnf-cache-{releasever}".format(releasever=releasever))
-                
-                # Log!
-                log("Analyzing a use case: {name} on {base_name} {release} ({number_pkgs} packages)".format(
-                        name=use_case["name"],
-                        base_name=bases[base_id]["name"],
-                        release=base_version,
-                        number_pkgs=len(packages)
-                ))
-
-                # Empty base?
-                number_of_base_pkgs = len(installs["bases"][base_install_id]["packages"])
-                empty_base = False
-                if number_of_base_pkgs == 0:
-                    empty_base = True
-
-                # Do the installation
-                installed_packages = _install_packages(installroot=base_installroot,
-                                                       pkgs_to_install=packages,
-                                                       options=options,
-                                                       releasever=releasever,
-                                                       cachedir=cachedir,
-                                                       installing_base_image=False,
-                                                       empty_base=empty_base)
-
-                # Save
-                use_case_installation = {}
-                use_case_installation["id"] = use_case_install_id
-                use_case_installation["use_case_id"] = use_case_id
-                use_case_installation["base_id"] = base_id
-                use_case_installation["base_version"] = base_version
-                use_case_installation["packages"] = installed_packages
-                installs["use_cases"][use_case_install_id] = use_case_installation
-                
-    return installs
-
-def get_data(configs, installs):
-    
-    data = {}
-
-    data["timestamp"] = installs["timestamp"]
-
-    data["base_definitions"] = configs["bases"]
-    data["base_installs"] = installs["bases"]
-
-    data["use_case_definitions"] = configs["use_cases"]
-    data["use_case_installs"] = installs["use_cases"]
-
-    # Store base info in use_case_definitions
-    for _, use_case_definition in data["use_case_definitions"].items():
-        base_versions = []
-        base_versions_set = set()
-        base_ids = []
-        base_ids_set = set()
-        for base_id in use_case_definition["install_on"]:
-            base_ids_set.add(base_id.split(":")[0])
-            base_versions_set.add(base_id.split(":")[1])
-        base_versions = list(base_versions_set)
-        base_versions.sort()
-        base_ids = list(base_ids_set)
-        base_ids.sort()
-        use_case_definition["base_versions"] = base_versions
-        use_case_definition["base_ids"] = base_ids
-
-        base_ids = use_case_definition["install_on"]
-        base_names = {}
-        for base_id in base_ids:
-            base_name, base_version = base_id.split(":")
-            if base_version not in base_names:
-                base_names[base_version] = []            
-            base_names[base_version].append(base_name)
-        use_case_definition["base_names"] = base_names
-
-    # Data focused on use cases
-    use_cases = {}
-    for _, use_case_install in data["use_case_installs"].items():
-        use_case_definition_id = use_case_install["use_case_id"]
-        use_case_definition = data["use_case_definitions"][use_case_definition_id]
-    
-        base_install_id = "{base_id}:{base_version}".format(
-                base_id=use_case_install["base_id"],
-                base_version=use_case_install["base_version"])
-        base_install = data["base_installs"][base_install_id]
-        base_id = use_case_install["base_id"]
-        base_definition = data["base_definitions"][base_id]
-        base_version = base_install["base_version"]
-
-        # Define a new entity holding all info about use cases
-        use_case = {}
-
-        # Identity and names
-        use_case["id"] = use_case_install["id"]
-        use_case["name"] = use_case_definition["name"]
-        use_case["file_id"] = "{use_case_id}--{base_id}--{base_version}".format(
-                use_case_id=use_case_definition["id"],
-                base_id=base_definition["id"],
-                base_version=base_version)
-        use_case["definition_id"] = use_case_definition["id"]
-
-        # Related data
-        use_case["base_name"] = base_definition["name"]
-        use_case["base_version"] = base_version
-        use_case["base_id"] = base_install_id
-        use_case["base_definition_id"] = use_case_install["base_id"]
-
-        # Total install size
-        total_size = 0
-        for _,package in use_case_install["packages"].items():
-            total_size += package["size"]
-        use_case["total_size"] = total_size
-
-        # Total install size history TODO
-        size_history = None
-        use_case["size_history"] = size_history
-
-        # Packages
-        use_case["packages"] = use_case_install["packages"]
-        package_names = []
-        for _,package in use_case_install["packages"].items():
-            package_names.append(package["name"])
-        use_case["package_names"] = package_names
-        use_case["required_package_names"] = use_case_definition["packages"]
-
-        # Advanced packages
-        base_package_names = []
-        for _,package in base_install["packages"].items():
-            base_package_names.append(package["name"])
-        packages_in_base = list(set(base_package_names) & set(package_names))
-        packages_not_in_base = list(set(package_names) - set(packages_in_base))
-        packages_in_base.sort()
-        packages_not_in_base.sort()
-        use_case["packages_in_base"] = packages_in_base
-        use_case["packages_not_in_base"] = packages_not_in_base
-
-        use_cases[use_case["id"]] = use_case
-
-    data["use_cases"] = use_cases
-
-    # Data focued on bases
-    bases = {}
-    for _, base_install in data["base_installs"].items():
-        base_definition = data["base_definitions"][base_install["base_id"]]
-        base_version = base_install["base_version"]
-        
-        # Define a new entity holding all the info about bases
-        base = {}
-        base["id"] = base_install["id"]
-        base["name"] = base_definition["name"]
-        base["version"] = base_version
-        base["file_id"] = "{id}--{version}".format(
-                id=base_definition["id"],
-                version=base_version)
-        base["definition_id"] = base_definition["id"]
-
-        # Related data
-        use_case_ids = []
-        for _,use_case in data["use_cases"].items():
-            if use_case["base_id"] == base["id"]:
-                use_case_ids.append(use_case["id"])
-        base["use_case_ids"] = use_case_ids
-
-        # Total install size
-        total_size = 0
-        for _,package in base_install["packages"].items():
-            total_size += package["size"]
-        base["total_size"] = total_size
-
-        # Total install size history TODO
-        size_history = None
-        base["size_history"] = size_history
-
-        # Packages
-        base["packages"] = base_install["packages"]
-        package_names = []
-        for _,package in base_install["packages"].items():
-            package_names.append(package["name"])
-        base["package_names"] = package_names
-        base["required_package_names"] = \
-                base_definition["versions"][base_version]["packages"]
-
-        bases[base["id"]] = base
-        
-    data["bases"] = bases
-
-    return data
-
-
-def _add_title_to_use_case_svg(svg, title_data):
-
-    try:
-        title_code = """
-  <text
-     id="text7664"
-     y="19.384439"
-     x="9.9055691"
-     style="font-style:normal;font-weight:normal;font-size:12.49485588px;line-height:125%;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.3123714px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1"
-     xml:space="preserve"><tspan
-       style="font-size:12.00258541px;stroke-width:0.3123714px"
-       y="19.384439"
-       x="9.9055691"
-       id="tspan7666"><a href="{use_case_url}"><tspan
-   style="font-weight:bold;font-size:12.00258541px;fill:#0066ff;stroke-width:0.3123714px"
-   id="tspan7668">{use_case_name}</tspan></a> on <a href="{base_url}"><tspan
-   style="font-weight:bold;font-size:12.00258541px;fill:#0066ff;stroke-width:0.3123714px"
-   id="tspan7676">{base_name}</tspan></a> (Fedora {version})</tspan></text>
-  <text
-     xml:space="preserve"
-     style="font-style:normal;font-weight:normal;font-size:11.85299969px;line-height:125%;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.296325px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1"
-     x="9.9365959"
-     y="58.520042"
-     id="text8235"><tspan
-       id="tspan8237"
-       x="9.9365959"
-       y="58.520042"
-       style="stroke-width:0.296325px;font-size:10.80232704px;fill:#666666">Generated by <a href="./"><tspan
-   style="fill:#3771c8"
-   id="tspan8255">Feedback Pipeline</tspan></a></tspan></text>
-  <text
-     xml:space="preserve"
-     style="font-style:normal;font-weight:normal;font-size:12.49485588px;line-height:125%;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.3123714px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1"
-     x="9.9055691"
-     y="38.974205"
-     id="text8245"><a href="{expand_collapse_url}"><tspan
-       id="tspan8247"
-       x="9.9055691"
-       y="38.974205"
-       style="font-weight:normal;font-size:12.00258541px;fill:#0066ff;stroke-width:0.3123714px">{expand_collapse_text}</tspan></a></text>
-""".format(
-        use_case_url=title_data["use_case_url"],
-        use_case_name=title_data["use_case_name"],
-        base_url=title_data["base_url"],
-        base_name=title_data["base_name"],
-        version=title_data["version"],
-        expand_collapse_url=title_data["expand_collapse_url"],
-        expand_collapse_text=title_data["expand_collapse_text"]
-    )
-
-    except KeyError:
-        return svg
-
-    return svg.split("</svg>")[0] + title_code + "\n</svg>\n"
-
-
-def _add_title_to_base_svg(svg, title_data):
-
-    try:
-        title_code = """
-  <text
-     
-     id="text7664"
-     y="19.384439"
-     x="9.9055691"
-     style="font-style:normal;font-weight:normal;font-size:12.49485588px;line-height:125%;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.3123714px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1"
-     xml:space="preserve"><tspan
-       style="font-size:12.00258541px;stroke-width:0.3123714px"
-       y="19.384439"
-       x="9.9055691"
-       id="tspan7666"
-       ><a href="{base_url}"><tspan
-   style="font-weight:bold;font-size:12.00258541px;fill:#0066ff;stroke-width:0.3123714px"
-   id="tspan7668">{base_name}</tspan></a> (Fedora {version})</tspan></text>
-  <text
-     xml:space="preserve"
-     style="font-style:normal;font-weight:normal;font-size:11.85299969px;line-height:125%;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.296325px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1"
-     x="9.9365959"
-     y="36.915382"
-     id="text8235"><tspan
-       
-       id="tspan8237"
-       x="9.9365959"
-       y="36.915382"
-       style="font-size:10.80232716px;fill:#666666;stroke-width:0.296325px">Generated by <a href="./"><tspan
-   style="fill:#3771c8"
-   id="tspan8255">Feedback Pipeline</tspan></a></tspan></text>
-""".format(
-        base_url=title_data["base_url"],
-        base_name=title_data["base_name"],
-        version=title_data["version"]
-    )
-
-    except KeyError:
-        return svg
-
-    return svg.split("</svg>")[0] + title_code + "\n</svg>\n"
-
-
-def generate_graphs(data, output):
-    log("Generating {} base graphs".format(len(data["bases"])))
-
-    for _,base in data["bases"].items():
-        log("  {} ({})".format(base["name"], base["version"]))
-
-        graph = showme.compute_graph(base["packages"])
-        highlights = base["required_package_names"]
-        dot = showme.graph_to_dot(graph, sizes=True, highlights=highlights)
-        svg = showme.dot_to_graph_svg(dot)
-
-        filename = "graph--{file_id}.svg".format(file_id=base["file_id"])
-
-        title_data = {}
-        title_data["base_url"] = "report-base-definition--{file_id}.html".format(file_id=base["file_id"])
-        title_data["base_name"] = base["name"]
-        title_data["version"] = base["version"]
-
-        svg = _add_title_to_base_svg(svg, title_data)
-
-        with open(os.path.join(output, filename), "w") as file:
-            file.write(svg)
-
-    log("Generating {} use case graphs".format(len(data["use_cases"])))
-
-    for _,use_case in data["use_cases"].items():
-        log("  {} on {} ({})".format(
-                use_case["name"], use_case["base_name"], use_case["base_version"]))
-
-        # Full graph shows all packages in the installation
-        graph_full = showme.compute_graph(use_case["packages"])
-        
-        # Simple graph groups the base image into a single node
-        base_packages = data["bases"][use_case["base_id"]]["packages"]
-        group = showme.packages_to_group(use_case["base_name"], base_packages)
-        graph_simple = showme.compute_graph(use_case["packages"], [group])
-
-        highlights = use_case["required_package_names"]
-        dot_full = showme.graph_to_dot(graph_full, sizes=True, highlights=highlights)
-        dot_simple = showme.graph_to_dot(graph_simple, sizes=True, highlights=highlights)
-
-        svg_full = showme.dot_to_graph_svg(dot_full)
-        svg_simple = showme.dot_to_graph_svg(dot_simple)
-
-        filename_full = "graph--{file_id}.svg".format(file_id=use_case["file_id"])
-        filename_simple = "graph-simple--{file_id}.svg".format(file_id=use_case["file_id"])
-
-        title_data = {}
-        title_data["use_case_url"] = "report-use-case-definition--{use_case_id}--{version}.html".format(
-            use_case_id=use_case["definition_id"],
-            version=use_case["base_version"])
-        title_data["use_case_name"] = use_case["name"]
-        title_data["base_url"] = "report-base-definition--{file_id}.html".format(file_id=data["bases"][use_case["base_id"]]["file_id"])
-        title_data["base_name"] = use_case["base_name"]
-        title_data["version"] = use_case["base_version"]
-        title_data["expand_collapse_url"] = filename_simple
-        title_data["expand_collapse_text"] = "Collapse base image"
-
-        svg_full = _add_title_to_use_case_svg(svg_full, title_data)
-
-        title_data["expand_collapse_url"] = filename_full
-        title_data["expand_collapse_text"] = "Expand base image"
-
-        svg_simple = _add_title_to_use_case_svg(svg_simple, title_data)
-
-        
-
-        with open(os.path.join(output, filename_full), "w") as file:
-            file.write(svg_full)
-
-        with open(os.path.join(output, filename_simple), "w") as file:
-            file.write(svg_simple)
-
-
-        
-
-def generate_reports_by_base(data, output):
-    log("Generating reports: Bases with use cases")
-
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-
-    for _,base in data["bases"].items():
-        base_report_data = {
-            "name": base["name"],
-            "version": base["version"],
-            "size": showme.size(base["total_size"]),
-            "packages": base["package_names"],
-            "file_id": base["file_id"]
-        }
-
-        pkg_sizes = {}
-        pkg_sizes_num = {}
-
-        # Get sizes of all packags in the base
-        for _,pkg in base["packages"].items():
-            if pkg["name"] not in pkg_sizes:
-                pkg_sizes[pkg["name"]] = showme.size(pkg["size"])
-                pkg_sizes_num[pkg["name"]] = pkg["size"]
-                
-
-        use_case_data = []
-        for use_case_id in base["use_case_ids"]:
-            use_case = data["use_cases"][use_case_id]
-
-            # Get sizes of all other packages
-            for _,pkg in use_case["packages"].items():
-                if pkg["name"] not in pkg_sizes:
-                    pkg_sizes[pkg["name"]] = showme.size(pkg["size"])
-                    pkg_sizes_num[pkg["name"]] = pkg["size"]
-
-            use_case_report_data = {
-                "name": use_case["name"],
-                "definition_id": use_case["definition_id"],
-                "size": showme.size(use_case["total_size"]),
-                "pkgs_in_base": use_case["packages_in_base"],
-                "pkgs_not_in_base": use_case["packages_not_in_base"],
-                "required_pkgs": use_case["required_package_names"],
-                "packages": use_case["packages"],
-                "file_id": use_case["file_id"]
-            } 
-
-            use_case_data.append(use_case_report_data)
-        
-        # Sort use cases by name
-        use_case_data_sorted = sorted(use_case_data, key=lambda k: k["name"])
-
-        extra_pkgs = []
-        for install_data in use_case_data_sorted:
-            extra_pkgs += install_data["pkgs_not_in_base"]
-        extra_pkgs = list(set(extra_pkgs))
-        extra_pkgs.sort()
-
-        table_report_template = template_env.get_template("report_by_base.html")
-        table_report = table_report_template.render(
-                base=base_report_data,
-                images=use_case_data_sorted,
-                extra_pkgs=extra_pkgs,
-                pkg_sizes=pkg_sizes,
-                pkg_sizes_num=pkg_sizes_num,
-                timestamp=data["timestamp"])
-
-        filename = "report-by-base--{file_id}.html".format(
-                file_id=base["file_id"])
-
-        with open(os.path.join(output, filename), "w") as file:
-            file.write(table_report)
-
-
-def generate_reports_by_use_case(data, output):
-    log("Generating reports: Use cases on bases")
-
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-
-    for _,use_case_definition in data["use_case_definitions"].items():
-        for base_version in use_case_definition["base_versions"]:
-            use_case_report_data = {
-                "name": use_case_definition["name"],
-                "packages": use_case_definition["packages"],
-                "version": base_version
-            }
-
-            pkg_sizes = {}
-            pkg_sizes_num = {}
-
-
-            other_install_data = []
-            for base_definition_id in use_case_definition["base_ids"]:
-                base_id = "{base_definition_id}:{base_version}".format(
-                        base_definition_id=base_definition_id,
-                        base_version=base_version)
-                use_case_id = "{use_case_definition_id}:{base_id}".format(
-                        use_case_definition_id=use_case_definition["id"],
-                        base_id=base_id)
-
-                base = data["bases"][base_id]
-                use_case = data["use_cases"][use_case_id]
-
-                # Get sizes of all packags in the base
-                for _,pkg in use_case["packages"].items():
-                    if pkg["name"] not in pkg_sizes:
-                        pkg_sizes[pkg["name"]] = showme.size(pkg["size"])
-                        pkg_sizes_num[pkg["name"]] = pkg["size"]
-
-                required_package_names = use_case["required_package_names"]
-                all_package_names = use_case["package_names"]
-                dependencies = list(set(all_package_names) - set(required_package_names))
-                dependencies.sort()
-
-                base_report_data = {
-                    "name": base["name"],
-                    "definition_id": base["definition_id"],
-                    "size": showme.size(use_case["total_size"]),
-                    "required_pkgs": required_package_names,
-                    "pkgs_in_base": use_case["packages_in_base"],
-                    "dependencies": dependencies,
-                    "packages": use_case["packages"],
-                    "file_id": use_case["file_id"]
-                }
-                other_install_data.append(base_report_data)
-
-            extra_pkgs = []
-            for install_data in other_install_data:
-                extra_pkgs += install_data["dependencies"]
-            extra_pkgs = list(set(extra_pkgs))
-            extra_pkgs.sort()
-
-            table_report_template = template_env.get_template("report_by_use_case.html")
-            table_report = table_report_template.render(
-                    base=use_case_report_data,
-                    images=other_install_data,
-                    extra_pkgs=extra_pkgs,
-                    pkg_sizes=pkg_sizes,
-                    pkg_sizes_num=pkg_sizes_num,
-                    timestamp=data["timestamp"])
-
-            file_id = "{use_case}--{version}".format(
-                    use_case=use_case_definition["id"],
-                    version=base_version)
-            filename = "report-by-use-case--{file_id}.html".format(
-                    file_id=file_id)
-
-            with open(os.path.join(output, filename), "w") as file:
-                file.write(table_report)
-
-
-def generate_reports_bases_releases(data, output):
-    log("Generating reports: Bases by releases")
-
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-
-    for _,base_definition in data["base_definitions"].items():
-        report_data = []
-        for base_version in base_definition["versions"]:
-            base_id = "{base_definition_id}:{base_version}".format(
-                    base_definition_id=base_definition["id"],
-                    base_version=base_version)
-            base = data["bases"][base_id]
-    
-            base_report_data = {
-                "version": base_version,
-                "size": showme.size(base["total_size"]),
-                "required_pkgs": base["required_package_names"],
-                "packages": base["packages"],
-                "file_id": base["file_id"]
-            }
-            report_data.append(base_report_data)
-
-        all_packages = []
-        for base_report_data in report_data:
-            all_packages += base_report_data["packages"].keys()
-        all_packages = list(set(all_packages))
-        all_packages.sort()
-
-        table_report_template = template_env.get_template("report_base_releases.html")
-        table_report = table_report_template.render(
-                report_data=report_data,
-                all_packages=all_packages,
-                base_name=base_definition["name"],
-                size_function=showme.size,
-                timestamp=data["timestamp"])
-
-        filename = "report-base-releases--{base_id}.html".format(
-            base_id=base_definition["id"])
-
-        with open(os.path.join(output, filename), "w") as file:
-            file.write(table_report)
-
-
-def load_historic_data(data, output):
-
-    log("Loading historic size data from disk")
-    directory = os.path.join(output, "history")
-
-    all_filenames = os.listdir(directory)
-    all_filenames.sort()
-
-    historic_data = {}
-
-    #FIXME: This only needs to load a limited number of data.
-    #       But for now it's ok, we don't have that much. (This is going to end up badly, right?)
-
-    for filename in all_filenames:
-        with open(os.path.join(directory, filename), "r") as file:
-            document = json.load(file)
-
-            date = datetime.datetime.strptime(document["timestamp"],"%d/%m/%Y %H:%M")
-            key = datetime.datetime.strftime(date, "%Y-%m-%d-%H%M")
-
-            historic_data[key] = document
-
-    return historic_data
-
-
-def get_historic_chart_data(historic_data, data):
-    log("Generating historic size chart data")
-
-    for base_id, base in data["bases"].items():
-        size_history = {}
-        for timestamp in sorted(historic_data):
-            historic_data_instance = historic_data[timestamp]
-            try:
-                size = historic_data_instance["bases"][base_id]["total_size"]
-            except KeyError:
-                continue
-
-            size_history[timestamp] = size
-        
-        base["size_history"] = size_history
-
-    for use_case_id, use_case in data["use_cases"].items():
-        size_history = {}
-        for timestamp in sorted(historic_data):
-            historic_data_instance = historic_data[timestamp]
-            try:
-                size = historic_data_instance["use_cases"][use_case_id]["total_size"]
-            except KeyError:
-                continue
-
-            size_history[timestamp] = size
-
-        use_case["size_history"] = size_history
-
-def generate_reports_bases_definitions(data, output):
-    log("Generating reports: Base definitions")
-
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-
-    for base_id, base_definition in data["base_definitions"].items():
-        for base_version in base_definition["versions"]:
-            report_template = template_env.get_template("report_base_definition.html")
-            base = data["bases"][base_id+":"+base_version]
-
-            chart_data_x = []
-            chart_data_y = []
-            for x, y in base["size_history"].items():
-                chart_data_x.append(x)
-                chart_data_y.append(round(y / 1024 / 1024,2))
-            report = report_template.render(
-                    base_definition=base_definition,
-                    base_version=base_version,
-                    base=base,
-                    chart_data_x=chart_data_x,
-                    chart_data_y=chart_data_y,
-                    size_function=showme.size,
-                    data=data)
-
-            filename = "report-base-definition--{base_id}--{base_version}.html".format(
-                base_id=base_definition["id"], base_version=base_version)
-
-            with open(os.path.join(output, filename), "w") as file:
-                file.write(report)
-
-
-def generate_reports_use_cases_definitions(data, output):
-    log("Generating reports: use_case definitions")
-
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-
-    for _,use_case_definition in data["use_case_definitions"].items():
-        
-        for base_version, base_ids in use_case_definition["base_names"].items():
-
-            use_cases = []
-            for base_id in base_ids:
-                use_case = data["use_cases"]["{use_case}:{base}:{version}".format(
-                    use_case=use_case_definition["id"],
-                    base=base_id,
-                    version=base_version
-                )]
-                use_cases.append(use_case)
-            
-            graph_timestamps_set = set()
-            for use_case in use_cases:
-                for timestamp, size in use_case["size_history"].items():
-                    graph_timestamps_set.add(timestamp)
-            
-            graph_sizes = {}
-            graph_timestamps = sorted(list(graph_timestamps_set))
-            for timestamp in graph_timestamps:
-                for use_case in use_cases:
-                    base_id = use_case["base_definition_id"]
-
-                    if base_id not in graph_sizes:
-                        graph_sizes[base_id] = []
-
-                    # The size is string because I need to be able to carry either
-                    # the number or "null" that will be put into Javascript code
-                    # in the jinja2 template. I know it's horrible.
-                    if timestamp in use_case["size_history"]:
-                        size_bytes = use_case["size_history"][timestamp]
-                        size = str(round(size_bytes / 1024 / 1024,2))
-                    else:
-                        size = "null"
-
-                    graph_sizes[base_id].append(size)
-            
-            report_template = template_env.get_template("report_use_case_definition.html")
-            report = report_template.render(
-                    use_case_definition=use_case_definition,
-                    base_version=base_version,
-                    graph_timestamps=graph_timestamps,
-                    graph_sizes=graph_sizes,
-                    data=data)
-
-            filename = "report-use-case-definition--{use_case_id}--{base_version}.html".format(
-                use_case_id=use_case_definition["id"],
-                base_version=base_version
-            )
-
-            with open(os.path.join(output, filename), "w") as file:
-                file.write(report)
-
-
-def generate_pages(data, output):
-    log("Generating common pages")
-
-    data = sort_definitions(copy.deepcopy(data))
-    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
-    template_env = jinja2.Environment(loader=template_loader)
-  
-    pages = [
-        # Template name                             # Output file
-        ("homepage.html",                           "index.html"),
-        ("view.html",                               "views.html"),
-        ("view_bases_definitions.html",             "view--bases-definitions.html"),
-        ("view_bases_compare_releases.html",        "view--bases-by-releases.html"),
-        ("view_use_cases_definitions.html",         "view--use-cases-definitions.html"),
-        ("view_use_cases_compare_bases.html",       "view--use-cases-on-bases.html"),
-        ("view_use_cases_compare_releases.html",    "view--use-cases-by-releases.html"),
-        ("view_use_cases_use_cases_by_base.html",   "view--bases-with-use-cases.html"),
-    ]
-
-
-    for template_name, output_file in pages:
-        template = template_env.get_template(template_name)
-        page = template.render(data=data)
-        with open(os.path.join(output, output_file), "w") as file:
-            file.write(page)
-    
-
-    src_static_dir = os.path.join("templates", "_static")
-    output_static_dir = os.path.join(output)
-    subprocess.run(["cp", "-R", src_static_dir, output_static_dir])
-
-
-
 
 def dump_data(path, data):
     with open(path, 'w') as file:
@@ -1102,64 +68,2490 @@ def dump_data(path, data):
 def load_data(path):
     with open(path, 'r') as file:
         data = json.load(file)
+    return data
+
+def size(num, suffix='B'):
+    for unit in ['','k','M','G']:
+        if abs(num) < 1024.0:
+            return "%3.1f %s%s" % (num, unit, suffix)
+        num /= 1024.0
+    return "%.1f %s%s" % (num, 'T', suffix)
+
+def load_settings():
+    settings = {}
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument("configs", help="Directory with YAML configuration files. Only files ending with '.yaml' are accepted.")
+    parser.add_argument("output", help="Directory to contain the output.")
+    args = parser.parse_args()
+
+    settings["configs"] = args.configs
+    settings["output"] = args.output
+
+    # FIXME: This is hardcorded, and it shouldn't be!
+    #settings["allowed_arches"] = ["armv7hl","aarch64","i686","ppc64le","s390x","x86_64"]
+    # FIXME Limiting arches for faster results during development
+    settings["allowed_arches"] = ["armv7hl","aarch64","ppc64le","s390x","x86_64"]
+    return(settings)
+
+
+
+
+###############################################################################
+### Loading user-provided configs #############################################
+###############################################################################
+
+# Configs:
+#   TYPE:         KEY:          ID:
+# - repo          repos         repo_id
+# - env           envs          env_id
+# - workload      workloads     workload_id
+# - label         labels        label_id
+# - view          views         view_id
+
+def _load_config_repo(document_id, document, settings):
+    config = {}
+    config["id"] = document_id
+
+    # Step 1: Mandatory fields
+    try:
+        # Name is an identifier for humans
+        config["name"] = str(document["data"]["name"])
+
+        # A short description, perhaps hinting the purpose
+        config["description"] = str(document["data"]["description"])
+
+        # Who maintains it? This is just a freeform string
+        # for humans to read. In Fedora, a FAS nick is recommended.
+        config["maintainer"] = str(document["data"]["maintainer"])
+
+        # Where does this repository come from?
+        # Right now, only Fedora repositories are supported,
+        # defined by their releasever.
+        config["source"] = {}
+
+        # Only Fedora repos supported at this time.
+        # Fedora release.
+        config["source"]["fedora_release"] = str(document["data"]["source"]["fedora_release"])
+
+        # List of architectures
+        config["source"]["architectures"] = []
+        for arch_raw in document["data"]["source"]["architectures"]:
+            arch = str(arch_raw)
+            if arch not in settings["allowed_arches"]:
+                err_log("Warning: {file}.yaml lists an invalid architecture: {arch}. Ignoring.".format(
+                    file=document_id,
+                    arch=arch))
+                continue
+            config["source"]["architectures"].append(str(arch))
+    except KeyError:
+        raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+    
+    # Step 2: Optional fields
+    # none here
+
+    return config
+
+
+def _load_config_env(document_id, document, settings):
+    config = {}
+    config["id"] = document_id
+
+    # Step 1: Mandatory fields
+    try:
+        # Name is an identifier for humans
+        config["name"] = str(document["data"]["name"])
+
+        # A short description, perhaps hinting the purpose
+        config["description"] = str(document["data"]["description"])
+
+        # Who maintains it? This is just a freeform string
+        # for humans to read. In Fedora, a FAS nick is recommended.
+        config["maintainer"] = str(document["data"]["maintainer"])
+
+        # Different instances of the environment, one per repository.
+        config["repositories"] = []
+        for repo in document["data"]["repositories"]:
+            config["repositories"].append(str(repo))
+        
+        # Packages defining this environment.
+        # This list includes packages for all
+        # architectures — that's the one to use by default.
+        config["packages"] = []
+        for pkg in document["data"]["packages"]:
+            config["packages"].append(str(pkg))
+        
+        # Labels connect things together.
+        # Workloads get installed in environments with the same label.
+        # They also get included in views with the same label.
+        config["labels"] = []
+        for repo in document["data"]["labels"]:
+            config["labels"].append(str(repo))
+
+    except KeyError:
+        raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+
+    # Step 2: Optional fields
+
+    # Architecture-specific packages.
+    config["arch_packages"] = {}
+    for arch in settings["allowed_arches"]:
+        config["arch_packages"][arch] = []
+    if "arch_packages" in document["data"]:
+        for arch, pkgs in document["data"]["arch_packages"].items():
+            if arch not in settings["allowed_arches"]:
+                err_log("Warning: {file}.yaml lists an invalid architecture: {arch}. Ignoring.".format(
+                    file=document_id,
+                    arch=arch
+                ))
+                continue
+            for pkg_raw in pkgs:
+                pkg = str(pkg_raw)
+                config["arch_packages"][arch].append(pkg)
+    
+    # Extra installation options.
+    # The following are now supported:
+    # - "include-docs" - include documentation packages
+    # - "include-weak-deps" - automatically pull in "recommends" weak dependencies
+    config["options"] = []
+    if "options" in document["data"]:
+        if "include-docs" in document["data"]["options"]:
+            config["options"].append("include-docs")
+        if "include-weak-deps" in document["data"]["options"]:
+            config["options"].append("include-weak-deps")
+
+    return config
+
+
+def _load_config_workload(document_id, document, settings):
+    config = {}
+    config["id"] = document_id
+
+    # Step 1: Mandatory fields
+    try:
+        # Name is an identifier for humans
+        config["name"] = str(document["data"]["name"])
+
+        # A short description, perhaps hinting the purpose
+        config["description"] = str(document["data"]["description"])
+
+        # Who maintains it? This is just a freeform string
+        # for humans to read. In Fedora, a FAS nick is recommended.
+        config["maintainer"] = str(document["data"]["maintainer"])
+
+        # Packages defining this workload.
+        # This list includes packages for all
+        # architectures — that's the one to use by default.
+        config["packages"] = []
+        for pkg in document["data"]["packages"]:
+            config["packages"].append(str(pkg))
+        
+        # Labels connect things together.
+        # Workloads get installed in environments with the same label.
+        # They also get included in views with the same label.
+        config["labels"] = []
+        for repo in document["data"]["labels"]:
+            config["labels"].append(str(repo))
+
+    except KeyError:
+        raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+
+    # Step 2: Optional fields
+
+    # Architecture-specific packages.
+    config["arch_packages"] = {}
+    for arch in settings["allowed_arches"]:
+        config["arch_packages"][arch] = []
+    if "arch_packages" in document["data"]:
+        for arch, pkgs in document["data"]["arch_packages"].items():
+            if arch not in settings["allowed_arches"]:
+                err_log("Error: {file}.yaml lists an invalid architecture: {arch}. Ignoring.".format(
+                    file=document_id,
+                    arch=arch
+                ))
+                continue
+            for pkg_raw in pkgs:
+                pkg = str(pkg_raw)
+                config["arch_packages"][arch].append(pkg)
+    
+    # Extra installation options.
+    # The following are now supported:
+    # - "include-docs" - include documentation packages
+    # - "include-weak-deps" - automatically pull in "recommends" weak dependencies
+    config["options"] = []
+    if "options" in document["data"]:
+        if "include-docs" in document["data"]["options"]:
+            config["options"].append("include-docs")
+        if "include-weak-deps" in document["data"]["options"]:
+            config["options"].append("include-weak-deps")
+
+    return config
+
+
+def _load_config_label(document_id, document, settings):
+    config = {}
+    config["id"] = document_id
+
+    # Step 1: Mandatory fields
+    try:
+        # Name is an identifier for humans
+        config["name"] = str(document["data"]["name"])
+
+        # A short description, perhaps hinting the purpose
+        config["description"] = str(document["data"]["description"])
+
+        # Who maintains it? This is just a freeform string
+        # for humans to read. In Fedora, a FAS nick is recommended.
+        config["maintainer"] = str(document["data"]["maintainer"])
+
+    except KeyError:
+        raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+
+    # Step 2: Optional fields
+    # none here
+
+    return config
+
+
+def _load_config_view(document_id, document, settings):
+    config = {}
+    config["id"] = document_id
+
+    # Step 1: Mandatory fields
+    try:
+        # Name is an identifier for humans
+        config["name"] = str(document["data"]["name"])
+
+        # A short description, perhaps hinting the purpose
+        config["description"] = str(document["data"]["description"])
+
+        # Who maintains it? This is just a freeform string
+        # for humans to read. In Fedora, a FAS nick is recommended.
+        config["maintainer"] = str(document["data"]["maintainer"])
+
+        # Labels connect things together.
+        # Workloads get installed in environments with the same label.
+        # They also get included in views with the same label.
+        config["labels"] = []
+        for repo in document["data"]["labels"]:
+            config["labels"].append(str(repo))
+
+    except KeyError:
+        raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+
+    # Step 2: Optional fields
+    # none here
+
+    return config
+
+
+def get_configs(settings):
+    log("")
+    log("###############################################################################")
+    log("### Loading user-provided configs #############################################")
+    log("###############################################################################")
+    log("")
+
+    directory = settings["configs"]
+
+    if "allowed_arches" not in settings:
+        err_log("System error: allowed_arches not configured")
+        raise SettingsError
+    
+    if not settings["allowed_arches"]:
+        err_log("System error: no allowed_arches not configured")
+        raise SettingsError
+
+    configs = {}
+
+    configs["repos"] = {}
+    configs["envs"] = {}
+    configs["workloads"] = {}
+    configs["labels"] = {}
+    configs["views"] = {}
+
+    for yml_file in os.listdir(directory):
+        # Only accept yaml files
+        if not yml_file.endswith(".yaml"):
+            continue
+        
+        document_id = yml_file.split(".yaml")[0]
+
+        try:
+            with open(os.path.join(directory, yml_file), "r") as file:
+                # Safely load the config
+                try:
+                    document = yaml.safe_load(file)
+                except yaml.YAMLError as err:
+                    raise ConfigError("Error loading a config '{filename}': {err}".format(
+                                filename=yml_file,
+                                err=err))
+                
+                # Only accept yaml files stating their purpose!
+                if not ("document" in document and "version" in document):
+                    raise ConfigError("Error: {file} is invalid.".format(file=yml_file))
+
+
+                # === Case: Repository config ===
+                if document["document"] == "feeback-pipeline-repository":
+                    configs["repos"][document_id] = _load_config_repo(document_id, document, settings)
+
+                # === Case: Environment config ===
+                if document["document"] == "feeback-pipeline-environment":
+                    configs["envs"][document_id] = _load_config_env(document_id, document, settings)
+
+                # === Case: Workload config ===
+                if document["document"] == "feeback-pipeline-workload":
+                    configs["workloads"][document_id] = _load_config_workload(document_id, document, settings)
+                
+                # === Case: Label config ===
+                if document["document"] == "feeback-pipeline-label":
+                    configs["labels"][document_id] = _load_config_label(document_id, document, settings)
+
+                # === Case: View config ===
+                if document["document"] == "feeback-pipeline-view":
+                    configs["views"][document_id] = _load_config_view(document_id, document, settings)
+
+
+        except ConfigError as err:
+            err_log("Config load error: {err}".format(err=err))
+            continue
+    
+    log("Done!  Loaded:")
+    log("  - {} repositories".format(len(configs["repos"])))
+    log("  - {} environments".format(len(configs["envs"])))
+    log("  - {} workloads".format(len(configs["workloads"])))
+    log("  - {} labels".format(len(configs["labels"])))
+    log("  - {} views".format(len(configs["views"])))
+    log("")
+
+    return configs
+
+
+
+###############################################################################
+### Analyzing stuff! ##########################################################
+###############################################################################
+
+# Configs:
+#   TYPE:           KEY:          ID:
+# - repo            repos         repo_id
+# - env_conf        envs          env_id
+# - workload_conf   workloads     workload_id
+# - label           labels        label_id
+# - conf_view       views         view_id
+#
+# Data:
+#   TYPE:         KEY:                 ID:
+# - pkg           pkgs/repo_id/arch    NEVR
+# - env           envs                 env_id:repo_id:arch_id
+# - workload      workloads            workload_id:env_id:repo_id:arch_id
+# - view          views                view_id:repo_id:arch_id
+#
+# tmp contents:
+# - dnf_cachedir-{repo}-{arch}
+# - dnf_generic_installroot-{repo}-{arch}
+# - dnf_env_installroot-{env_conf}-{repo}-{arch}
+#
+#
+
+def _analyze_pkgs(tmp, repo, arch):
+    log("Analyzing pkgs for {repo_name} ({repo_id}) {arch}".format(
+            repo_name=repo["name"],
+            repo_id=repo["id"],
+            arch=arch
+        ))
+    
+    with dnf.Base() as base:
+
+        # Local DNF cache
+        cachedir_name = "dnf_cachedir-{repo}-{arch}".format(
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.cachedir = os.path.join(tmp, cachedir_name)
+
+        # Generic installroot
+        root_name = "dnf_generic_installroot-{repo}-{arch}".format(
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.installroot = os.path.join(tmp, root_name)
+
+        # Architecture
+        base.conf.arch = arch
+        base.conf.ignorearch = True
+
+        # Repository
+        base.conf.substitutions['releasever'] = repo["source"]["fedora_release"]
+
+        # Load repos
+        log("  Loading repos...")
+        base.read_all_repos()
+        # This sometimes fails, so let's try at least N times
+        # before totally giving up!
+        MAX_TRIES = 10
+        attempts = 0
+        success = False
+        while attempts < MAX_TRIES:
+            try:
+                base.fill_sack(load_system_repo=False)
+                success = True
+                break
+            except dnf.exceptions.RepoError as err:
+                attempts +=1
+                log("  Failed to download repodata. Trying again!")
+        if not success:
+            err = "Failed to download repodata while analyzing repo '{repo_name} ({repo_id}) {arch}".format(
+            repo_name=repo["name"],
+            repo_id=repo["id"],
+            arch=arch
+            )
+            err_log(err)
+            raise RepoDownloadError(err)
+
+        # DNF query
+        query = base.sack.query
+
+        # Get all packages
+        all_pkgs_set = set(query())
+        pkgs = {}
+        for pkg_object in all_pkgs_set:
+            pkg_nevr = "{name}-{evr}".format(name=pkg_object.name, evr=pkg_object.evr)
+            pkg = {}
+            pkg["id"] = pkg_nevr
+            pkg["name"] = pkg_object.name
+            pkg["evr"] = pkg_object.evr
+            pkg["arch"] = pkg_object.arch
+            pkg["installsize"] = pkg_object.installsize
+            pkg["description"] = pkg_object.description
+            #pkg["provides"] = pkg_object.provides
+            #pkg["requires"] = pkg_object.requires
+            #pkg["recommends"] = pkg_object.recommends
+            #pkg["suggests"] = pkg_object.suggests
+            pkg["summary"] = pkg_object.summary
+            pkgs[pkg_nevr] = pkg
+        
+        log("  Done!  ({pkg_count} packages in total)".format(
+            pkg_count=len(pkgs)
+        ))
+        log("")
+
+    return pkgs
+
+
+def _analyze_env(tmp, env_conf, repo, arch):
+    env = {}
+    
+    env["env_conf_id"] = env_conf["id"]
+    env["pkg_ids"] = []
+    env["repo_id"] = repo["id"]
+    env["arch"] = arch
+
+    env["errors"] = {}
+    env["errors"]["non_existing_pkgs"] = []
+
+    env["succeeded"] = True
+
+    with dnf.Base() as base:
+
+        # Local DNF cache
+        cachedir_name = "dnf_cachedir-{repo}-{arch}".format(
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.cachedir = os.path.join(tmp, cachedir_name)
+
+        # Environment installroot
+        root_name = "dnf_env_installroot-{env_conf}-{repo}-{arch}".format(
+            env_conf=env_conf["id"],
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.installroot = os.path.join(tmp, root_name)
+
+        # Architecture
+        base.conf.arch = arch
+        base.conf.ignorearch = True
+
+        # Repository
+        base.conf.substitutions['releasever'] = repo["source"]["fedora_release"]
+
+        # Additional DNF Settings
+        base.conf.tsflags.append('justdb')
+
+        # Environment config
+        if "include-weak-deps" not in env_conf["options"]:
+            base.conf.install_weak_deps = False
+        if "include-docs" not in env_conf["options"]:
+            base.conf.tsflags.append('nodocs')
+
+        # Load repos
+        log("  Loading repos...")
+        base.read_all_repos()
+        # This sometimes fails, so let's try at least N times
+        # before totally giving up!
+        MAX_TRIES = 10
+        attempts = 0
+        success = False
+        while attempts < MAX_TRIES:
+            try:
+                base.fill_sack(load_system_repo=False)
+                success = True
+                break
+            except dnf.exceptions.RepoError as err:
+                attempts +=1
+                log("  Failed to download repodata. Trying again!")
+        if not success:
+            err = "Failed to download repodata while analyzing environment '{env_conf}' from '{repo}' {arch}:".format(
+                env_conf=env_conf["id"],
+                repo=repo["id"],
+                arch=arch
+            )
+            err_log(err)
+            raise RepoDownloadError(err)
+
+
+        # Packages
+        log("  Adding packages...")
+        for pkg in env_conf["packages"]:
+            try:
+                base.install(pkg)
+            except dnf.exceptions.MarkingError:
+                env["errors"]["non_existing_pkgs"].append(pkg)
+                continue
+
+        # Architecture-specific packages
+        for pkg in env_conf["arch_packages"][arch]:
+            try:
+                base.install(pkg)
+            except dnf.exceptions.MarkingError:
+                env["errors"]["non_existing_pkgs"].append(pkg)
+                continue
+        
+        # Resolve dependencies
+        log("  Resolving dependencies...")
+        base.resolve()
+
+        # Write the result into RPMDB.
+        # The transaction needs us to download all the packages. :(
+        # So let's do that to make it happy.
+        log("  Downloading packages...")
+        base.download_packages(base.transaction.install_set)
+        log("  Running DNF transaction, writing RPMDB...")
+        try:
+            base.do_transaction()
+        except dnf.exceptions.TransactionCheckError as err:
+            err_log("Failed to analyze environment '{env_conf}' from '{repo}' {arch}:".format(
+                    env_conf=env_conf["id"],
+                    repo=repo["id"],
+                    arch=arch
+                ))
+            err_log("  - {err}".format(err=err))
+            env["succeeded"] = False
+            return env
+
+        # DNF Query
+        log("  Creating a DNF Query object...")
+        query = base.sack.query().filterm(pkg=base.transaction.install_set)
+
+        for pkg in query:
+            pkg_id = "{name}-{evr}".format(
+                name=pkg.name,
+                evr=pkg.evr
+            )
+            env["pkg_ids"].append(pkg_id)
+        
+        log("  Done!  ({pkg_count} packages in total)".format(
+            pkg_count=len(env["pkg_ids"])
+        ))
+        log("")
+    
+    return env
+
+
+def _analyze_envs(tmp, configs):
+    envs = {}
+
+    # Look at all env configs...
+    for env_conf_id, env_conf in configs["envs"].items():
+        # For each of those, look at all repos it lists...
+        for repo_id in env_conf["repositories"]:
+            # And for each of the repo, look at all arches it supports.
+            repo = configs["repos"][repo_id]
+            for arch in repo["source"]["architectures"]:
+                # Now it has
+                #    all env confs *
+                #    repos each config lists *
+                #    archeas each repo supports
+                # Analyze all of that!
+                log("Analyzing {env_name} ({env_id}) from {repo_name} ({repo}) {arch}...".format(
+                    env_name=env_conf["name"],
+                    env_id=env_conf_id,
+                    repo_name=repo["name"],
+                    repo=repo_id,
+                    arch=arch
+                ))
+
+                env_id = "{env_conf_id}:{repo_id}:{arch}".format(
+                    env_conf_id=env_conf_id,
+                    repo_id=repo_id,
+                    arch=arch
+                )
+                envs[env_id] = _analyze_env(tmp, env_conf, repo, arch)
+                
+    
+    return envs
+
+
+def _analyze_workload(tmp, workload_conf, env_conf, repo, arch):
+    workload = {}
+
+    workload["workload_conf_id"] = workload_conf["id"]
+    workload["env_conf_id"] = env_conf["id"]
+    workload["repo_id"] = repo["id"]
+    workload["arch"] = arch
+
+    workload["pkg_env_ids"] = []
+    workload["pkg_added_ids"] = []
+
+    workload["errors"] = {}
+    workload["errors"]["non_existing_pkgs"] = []
+
+    workload["succeeded"] = True
+
+    with dnf.Base() as base:
+
+        # Local DNF cache
+        cachedir_name = "dnf_cachedir-{repo}-{arch}".format(
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.cachedir = os.path.join(tmp, cachedir_name)
+
+        # Environment installroot
+        # Since we're not writing anything into the installroot,
+        # let's just use the base image's installroot!
+        root_name = "dnf_env_installroot-{env_conf}-{repo}-{arch}".format(
+            env_conf=env_conf["id"],
+            repo=repo["id"],
+            arch=arch
+        )
+        base.conf.installroot = os.path.join(tmp, root_name)
+
+        # Architecture
+        base.conf.arch = arch
+        base.conf.ignorearch = True
+
+        # Repository
+        base.conf.substitutions['releasever'] = repo["source"]["fedora_release"]
+
+        # Environment config
+        if "include-weak-deps" not in workload_conf["options"]:
+            base.conf.install_weak_deps = False
+        if "include-docs" not in workload_conf["options"]:
+            base.conf.tsflags.append('nodocs')
+
+        # Load repos
+        log("  Loading repos...")
+        base.read_all_repos()
+
+        # Now I need to load the local RPMDB.
+        # However, if the environment is empty, it wasn't created, so I need to treat
+        # it differently. So let's check!
+        if len(env_conf["packages"]) or len(env_conf["arch_packages"][arch]):
+            # It's not empty! Load local data.
+            base.fill_sack(load_system_repo=True)
+        else:
+            # It's empty. Treat it like we're using an empty installroot.
+            # This sometimes fails, so let's try at least N times
+            # before totally giving up!
+            MAX_TRIES = 10
+            attempts = 0
+            success = False
+            while attempts < MAX_TRIES:
+                try:
+                    base.fill_sack(load_system_repo=False)
+                    success = True
+                    break
+                except dnf.exceptions.RepoError as err:
+                    attempts +=1
+                    log("  Failed to download repodata. Trying again!")
+            if not success:
+                err = "Failed to download repodata while analyzing workload '{workload_id} on '{env_id}' from '{repo}' {arch}...".format(
+                        workload_id=workload_conf_id,
+                        env_id=env_conf_id,
+                        repo_name=repo["name"],
+                        repo=repo_id,
+                        arch=arch)
+                err_log(err)
+                raise RepoDownloadError(err)
+        
+        # Packages
+        log("  Adding packages...")
+        for pkg in workload_conf["packages"]:
+            try:
+                base.install(pkg)
+            except dnf.exceptions.MarkingError:
+                workload["errors"]["non_existing_pkgs"].append(pkg)
+                continue
+
+        # Architecture-specific packages
+        for pkg in workload_conf["arch_packages"][arch]:
+            try:
+                base.install(pkg)
+            except dnf.exceptions.MarkingError:
+                workload["errors"]["non_existing_pkgs"].append(pkg)
+                continue
+
+        # Resolve dependencies
+        log("  Resolving dependencies...")
+        base.resolve()
+
+        # DNF Query
+        log("  Creating a DNF Query object...")
+        query_env = base.sack.query()
+        query_added = base.sack.query().filterm(pkg=base.transaction.install_set)
+        pkgs_env = set(query_env.installed())
+        pkgs_added = set(base.transaction.install_set)
+        pkgs_all = set.union(pkgs_env, pkgs_added)
+        query_all = base.sack.query().filterm(pkg=pkgs_all)
+        
+        for pkg in pkgs_env:
+            pkg_id = "{name}-{evr}".format(
+                name=pkg.name,
+                evr=pkg.evr
+            )
+            workload["pkg_env_ids"].append(pkg_id)
+        
+        for pkg in pkgs_added:
+            pkg_id = "{name}-{evr}".format(
+                name=pkg.name,
+                evr=pkg.evr
+            )
+            workload["pkg_added_ids"].append(pkg_id)
+        
+        pkg_env_count = len(workload["pkg_env_ids"])
+        pkg_added_count = len(workload["pkg_added_ids"])
+        log("  Done!  ({pkg_count} packages in total. That's {pkg_env_count} in the environment, and {pkg_added_count} added.)".format(
+            pkg_count=str(pkg_env_count + pkg_added_count),
+            pkg_env_count=pkg_env_count,
+            pkg_added_count=pkg_added_count
+        ))
+        log("")
+
+    return workload
+
+
+def _analyze_workloads(tmp, configs):
+    workloads = {}
+
+    # Here, I need to mix and match workloads & envs based on labels
+    workload_env_map = {}
+    # Look at all workload configs...
+    for workload_conf_id, workload_conf in configs["workloads"].items():
+        workload_env_map[workload_conf_id] = set()
+        # ... and all of their labels.
+        for label in workload_conf["labels"]:
+            # And for each label, find all env configs...
+            for env_conf_id, env_conf in configs["envs"].items():
+                # ... that also have the label.
+                if label in env_conf["labels"]:
+                    # And save those.
+                    workload_env_map[workload_conf_id].add(env_conf_id)
+    
+    # And now, look at all workload configs...
+    for workload_conf_id, workload_conf in configs["workloads"].items():
+        # ... and for each, look at all env configs it should be analyzed in.
+        for env_conf_id in workload_env_map[workload_conf_id]:
+            # Each of those envs can have multiple repos associated...
+            env_conf = configs["envs"][env_conf_id]
+            for repo_id in env_conf["repositories"]:
+                # ... and each repo probably has multiple architecture.
+                repo = configs["repos"][repo_id]
+                for arch in repo["source"]["architectures"]:
+
+                    # And now it has:
+                    #   all workload configs *
+                    #   all envs that match those *
+                    #   all repos of those envs *
+                    #   all arches of those repos.
+                    # That's a lot of stuff! Let's analyze all of that!
+                    log("Analyzing {workload_name} ({workload_id}) on {env_name} ({env_id}) from {repo_name} ({repo}) {arch}...".format(
+                        workload_name=workload_conf["name"],
+                        workload_id=workload_conf_id,
+                        env_name=env_conf["name"],
+                        env_id=env_conf_id,
+                        repo_name=repo["name"],
+                        repo=repo_id,
+                        arch=arch
+                    ))
+
+                    workload_id = "{workload_conf_id}:{env_conf_id}:{repo_id}:{arch}".format(
+                        workload_conf_id=workload_conf_id,
+                        env_conf_id=env_conf_id,
+                        repo_id=repo_id,
+                        arch=arch
+                    )
+
+                    workloads[workload_id] = _analyze_workload(tmp, workload_conf, env_conf, repo, arch)
+
+
+    return workloads
+
+
+def analyze_things(configs, settings):
+    log("")
+    log("###############################################################################")
+    log("### Analyzing stuff! ##########################################################")
+    log("###############################################################################")
+    log("")
+
+    data = {}
+
+    data["pkgs"] = {}
+    data["envs"] = {}
+    data["workloads"] = {}
+    data["views"] = {}
+
+    with tempfile.TemporaryDirectory() as tmp:
+
+        # FIXME temporary override
+        tmp = "/tmp/fixed-tmp"
+
+        # List of supported arches
+        all_arches = settings["allowed_arches"]
+
+        # Packages
+        log("")
+        log("=====  Analyzing Repos & Packages =====")
+        log("")
+        for _,repo in configs["repos"].items():
+            repo_id = repo["id"]
+            data["pkgs"][repo_id] = {}
+            for arch in repo["source"]["architectures"]:
+                data["pkgs"][repo_id][arch] = _analyze_pkgs(tmp, repo, arch)
+
+        # Environments
+        log("")
+        log("=====  Analyzing Environments =====")
+        log("")
+        data["envs"] = _analyze_envs(tmp, configs)
+
+        # Workloads
+        log("")
+        log("=====  Analyzing Workloads =====")
+        log("")
+        data["workloads"] = _analyze_workloads(tmp, configs)
+
 
     return data
 
 
-def sort_definitions(definitions):
-    definitions["base_definitions"] = sort_data(definitions["base_definitions"])
-    definitions["use_case_definitions"] = sort_data(definitions["use_case_definitions"])
-    return definitions
+###############################################################################
+### Query gives an easy access to the data! ###################################
+###############################################################################
+
+class Query():
+    def __init__(self, data, configs, settings):
+        self.data = data
+        self.configs = configs
+        self.settings = settings
+    
+    def size(self, num, suffix='B'):
+        for unit in ['','k','M','G']:
+            if abs(num) < 1024.0:
+                return "%3.1f %s%s" % (num, unit, suffix)
+            num /= 1024.0
+        return "%.1f %s%s" % (num, 'T', suffix)
+        
+
+    def workloads(self, workload_conf_id, env_conf_id, repo_id, arch, list_all=False, output_change=None):
+        # accepts none in any argument, and in those cases, answers for all instances
+
+        # It can output just one part of the id.
+        # That's useful to, for example, list all arches associated with a workload_conf_id
+        if output_change:
+            list_all = True
+            if output_change not in ["workload_conf_ids", "env_conf_ids", "repo_ids", "arches"]:
+                raise ValueError('output_change must be one of: "workload_conf_ids", "env_conf_ids", "repo_ids", "arches"')
+
+        matching_ids = set()
+
+        # list considered workload_conf_ids
+        if workload_conf_id:
+            workload_conf_ids = [workload_conf_id]
+        else:
+            workload_conf_ids = self.configs["workloads"].keys()
+
+        # list considered env_conf_ids
+        if env_conf_id:
+            env_conf_ids = [env_conf_id]
+        else:
+            env_conf_ids = self.configs["envs"].keys()
+        
+        # list considered repo_ids
+        if repo_id:
+            repo_ids = [repo_id]
+        else:
+            repo_ids = self.configs["repos"].keys()
+            
+        # list considered arches
+        if arch:
+            arches = [arch]
+        else:
+            arches = self.settings["allowed_arches"]
+        
+        # And now try looping through all of that, and return True on a first occurance
+        # This is a terrible amount of loops. But most cases will have just one item
+        # in most of those, anyway. No one is expected to run this method with
+        # a "None" for every argument!
+        for workload_conf_id in workload_conf_ids:
+            for env_conf_id in env_conf_ids:
+                for repo_id in repo_ids:
+                    for arch in arches:
+                        workload_id = "{workload_conf_id}:{env_conf_id}:{repo_id}:{arch}".format(
+                            workload_conf_id=workload_conf_id,
+                            env_conf_id=env_conf_id,
+                            repo_id=repo_id,
+                            arch=arch
+                        )
+                        if workload_id in self.data["workloads"].keys():
+                            if not list_all:
+                                return True
+                            if output_change:
+                                if output_change == "workload_conf_ids":
+                                    matching_ids.add(workload_conf_id)
+                                if output_change == "env_conf_ids":
+                                    matching_ids.add(env_conf_id)
+                                if output_change == "repo_ids":
+                                    matching_ids.add(repo_id)
+                                if output_change == "arches":
+                                    matching_ids.add(arch)
+                            else:
+                                matching_ids.add(workload_id)
+        
+        if not list_all:
+            return False
+        return sorted(list(matching_ids))
+    
+    def workloads_id(self, id, list_all=False, output_change=None):
+        # Accepts both env and workload ID, and returns workloads that match that
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.workloads(None, env_conf_id, repo_id, arch, list_all, output_change)
+        
+        # It's a workload! Why would you want that, anyway?!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.workloads(workload_conf_id, env_conf_id, repo_id, arch, list_all, output_change)
+        
+        raise ValueError("That seems to be an invalid ID!")
+
+    def envs(self, env_conf_id, repo_id, arch, list_all=False, output_change=None):
+        # accepts none in any argument, and in those cases, answers for all instances
+
+        # It can output just one part of the id.
+        # That's useful to, for example, list all arches associated with a workload_conf_id
+        if output_change:
+            list_all = True
+            if output_change not in ["env_conf_ids", "repo_ids", "arches"]:
+                raise ValueError('output_change must be one of: "env_conf_ids", "repo_ids", "arches"')
+        
+        matching_ids = set()
+
+        # list considered env_conf_ids
+        if env_conf_id:
+            env_conf_ids = [env_conf_id]
+        else:
+            env_conf_ids = self.configs["envs"].keys()
+        
+        # list considered repo_ids
+        if repo_id:
+            repo_ids = [repo_id]
+        else:
+            repo_ids = self.configs["repos"].keys()
+            
+        # list considered arches
+        if arch:
+            arches = [arch]
+        else:
+            arches = self.settings["allowed_arches"]
+        
+        # And now try looping through all of that, and return True on a first occurance
+        # This is a terrible amount of loops. But most cases will have just one item
+        # in most of those, anyway. No one is expected to run this method with
+        # a "None" for every argument!
+        for env_conf_id in env_conf_ids:
+            for repo_id in repo_ids:
+                for arch in arches:
+                    env_id = "{env_conf_id}:{repo_id}:{arch}".format(
+                        env_conf_id=env_conf_id,
+                        repo_id=repo_id,
+                        arch=arch
+                    )
+                    if env_id in self.data["envs"].keys():
+                        if not list_all:
+                            return True
+                        if output_change:
+                            if output_change == "env_conf_ids":
+                                matching_ids.add(env_conf_id)
+                            if output_change == "repo_ids":
+                                matching_ids.add(repo_id)
+                            if output_change == "arches":
+                                matching_ids.add(arch)
+                        else:
+                            matching_ids.add(env_id)
+        
+        # This means nothing has been found!
+        if not list_all:
+            return False
+        return sorted(list(matching_ids))
+    
+    def envs_id(self, id, list_all=False, output_change=None):
+        # Accepts both env and workload ID, and returns workloads that match that
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.envs(env_conf_id, repo_id, arch, list_all, output_change)
+        
+        # It's a workload!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.envs(env_conf_id, repo_id, arch, list_all, output_change)
+        
+        raise ValueError("That seems to be an invalid ID!")
+    
+    def workload_pkgs(self, workload_conf_id, env_conf_id, repo_id, arch):
+        # Warning: mixing repos and arches works, but might cause mess on the output
+
+        # Output is just a flat list. Extra fields will be added into each package:
+        # q_in          - set of workload_ids including this pkg
+        # q_required_in - set of workload_ids where this pkg is required (top-level)
+        # q_env_in      - set of workload_ids where this pkg is in env
+        # q_arch        - architecture
+        
+        # Step 1: get all the matching workloads!
+        workload_ids = self.workloads(workload_conf_id, env_conf_id, repo_id, arch, list_all=True)
+
+        # I'll need repo_ids and arches to access the packages
+        repo_ids = self.workloads(workload_conf_id, env_conf_id, repo_id, arch, output_change="repo_ids")
+        arches = self.workloads(workload_conf_id, env_conf_id, repo_id, arch, output_change="arches")
+
+        # Replicating the same structure as in data["pkgs"]
+        # That is: [repo_id][arch][pkg_id]
+        pkgs = {}
+        for repo_id in repo_ids:
+            pkgs[repo_id] = {}
+            for arch in arches:
+                pkgs[repo_id][arch] = {}
+
+        # Workloads are already paired with envs, repos, and arches
+        # (there is one for each combination)
+        for workload_id in workload_ids:
+            workload = self.data["workloads"][workload_id]
+            workload_arch = workload["arch"]
+            workload_repo_id = workload["repo_id"]
+            workload_conf_id = workload["workload_conf_id"]
+
+            # First, get all pkgs in the env
+            for pkg_id in workload["pkg_env_ids"]:
+
+                # Add it to the list if it's not there already.
+                # Create a copy since it's gonna be modified, and include only what's needed
+                if pkg_id not in pkgs[workload_repo_id][workload_arch]:
+                    pkg = self.data["pkgs"][workload_repo_id][workload_arch][pkg_id]
+                    pkgs[workload_repo_id][workload_arch][pkg_id] = {}
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["id"] = pkg_id
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["name"] = pkg["name"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["evr"] = pkg["evr"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["arch"] = pkg["arch"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["installsize"] = pkg["installsize"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["description"] = pkg["description"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["summary"] = pkg["summary"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_arch"] = workload_arch
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_in"] = set()
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"] = set()
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_env_in"] = set()
+                
+                # It's here, so add it
+                pkgs[workload_repo_id][workload_arch][pkg_id]["q_in"].add(workload_id)
+                # Browsing env packages, so add it
+                pkgs[workload_repo_id][workload_arch][pkg_id]["q_env_in"].add(workload_id)
+                # Is it required?
+                if pkg["name"] in self.configs["workloads"][workload_conf_id]["packages"]:
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"].add(workload_id)
+                if pkg["name"] in self.configs["workloads"][workload_conf_id]["arch_packages"][workload_arch]:
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"].add(workload_id)
+            
+            # Second, add all the other packages
+            for pkg_id in workload["pkg_added_ids"]:
+
+                # Add it to the list if it's not there already
+                # and initialize extra fields
+                if pkg_id not in pkgs[workload_repo_id][workload_arch]:
+                    pkg = self.data["pkgs"][workload_repo_id][workload_arch][pkg_id]
+                    pkgs[workload_repo_id][workload_arch][pkg_id] = {}
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["id"] = pkg_id
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["name"] = pkg["name"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["evr"] = pkg["evr"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["arch"] = pkg["arch"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["installsize"] = pkg["installsize"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["description"] = pkg["description"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["summary"] = pkg["summary"]
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_arch"] = workload_arch
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_in"] = set()
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"] = set()
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_env_in"] = set()
+                
+                # It's here, so add it
+                pkgs[workload_repo_id][workload_arch][pkg_id]["q_in"].add(workload_id)
+                # Not adding it to q_env_in
+                # Is it required?
+                if pkg["name"] in self.configs["workloads"][workload_conf_id]["packages"]:
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"].add(workload_id)
+                if pkg["name"] in self.configs["workloads"][workload_conf_id]["arch_packages"][workload_arch]:
+                    pkgs[workload_repo_id][workload_arch][pkg_id]["q_required_in"].add(workload_id)
+
+        # And now I just need to flatten that dict and return all packages as a list
+        final_pkg_list = []
+        for repo_id in repo_ids:
+            for arch in arches:
+                for pkg_id, pkg in pkgs[repo_id][arch].items():
+                    final_pkg_list.append(pkg)
+
+        # And sort them by nevr which is their ID
+        final_pkg_list_sorted = sorted(final_pkg_list, key=lambda k: k['id'])
+
+        return final_pkg_list_sorted
 
 
-def sort_data(data):
-    names = {}
-    for data_key, data_value in data.items():
-        data_dict = {}
-        data_dict[data_key] = data_value
-        names[data_value["name"]] = data_dict
-    sorted_list = sorted(names.items())
-    soted_data = []
-    for sorted_taple in sorted_list:
-        for _, item_value in sorted_taple[1].items():
-            soted_data.append(item_value)
-    return soted_data
+    def workload_pkgs_id(self, id):
+        # Accepts both env and workload ID, and returns pkgs for workloads that match
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.workload_pkgs(None, env_conf_id, repo_id, arch)
+        
+        # It's a workload!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.workload_pkgs(workload_conf_id, env_conf_id, repo_id, arch)
+        
+        raise ValueError("That seems to be an invalid ID!")
+    
+    def env_pkgs(self, env_conf_id, repo_id, arch):
+        # Warning: mixing repos and arches works, but might cause mess on the output
+
+        # Output is just a flat list. Extra fields will be added into each package:
+        # q_in          - set of env_ids including this pkg
+        # q_required_in - set of env_ids where this pkg is required (top-level)
+        # q_arch        - architecture
+
+        
+        # Step 1: get all the matching envs!
+        env_ids = self.envs(env_conf_id, repo_id, arch, list_all=True)
+
+        # I'll need repo_ids and arches to access the packages
+        repo_ids = self.envs(env_conf_id, repo_id, arch, output_change="repo_ids")
+        arches = self.envs(env_conf_id, repo_id, arch, output_change="arches")
+
+        # Replicating the same structure as in data["pkgs"]
+        # That is: [repo_id][arch][pkg_id]
+        pkgs = {}
+        for repo_id in repo_ids:
+            pkgs[repo_id] = {}
+            for arch in arches:
+                pkgs[repo_id][arch] = {}
+
+        # envs are already paired with repos, and arches
+        # (there is one for each combination)
+        for env_id in env_ids:
+            env = self.data["envs"][env_id]
+            env_arch = env["arch"]
+            env_repo_id = env["repo_id"]
+            env_conf_id = env["env_conf_id"]
+
+            for pkg_id in env["pkg_ids"]:
+
+                # Add it to the list if it's not there already.
+                # Create a copy since it's gonna be modified, and include only what's needed
+                if pkg_id not in pkgs[env_repo_id][env_arch]:
+                    pkg = self.data["pkgs"][env_repo_id][env_arch][pkg_id]
+                    pkgs[env_repo_id][env_arch][pkg_id] = {}
+                    pkgs[env_repo_id][env_arch][pkg_id]["id"] = pkg_id
+                    pkgs[env_repo_id][env_arch][pkg_id]["name"] = pkg["name"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["evr"] = pkg["evr"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["arch"] = pkg["arch"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["installsize"] = pkg["installsize"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["description"] = pkg["description"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["summary"] = pkg["summary"]
+                    pkgs[env_repo_id][env_arch][pkg_id]["q_arch"] = env_arch
+                    pkgs[env_repo_id][env_arch][pkg_id]["q_in"] = set()
+                    pkgs[env_repo_id][env_arch][pkg_id]["q_required_in"] = set()
+                
+                # It's here, so add it
+                pkgs[env_repo_id][env_arch][pkg_id]["q_in"].add(env_id)
+                # Is it required?
+                if pkg["name"] in self.configs["envs"][env_conf_id]["packages"]:
+                    pkgs[env_repo_id][env_arch][pkg_id]["q_required_in"].add(env_id)
+                if pkg["name"] in self.configs["envs"][env_conf_id]["arch_packages"][env_arch]:
+                    pkgs[env_repo_id][env_arch][pkg_id]["q_required_in"].add(env_id)
+
+        # And now I just need to flatten that dict and return all packages as a list
+        final_pkg_list = []
+        for repo_id in repo_ids:
+            for arch in arches:
+                for pkg_id, pkg in pkgs[repo_id][arch].items():
+                    final_pkg_list.append(pkg)
+
+        # And sort them by nevr which is their ID
+        final_pkg_list_sorted = sorted(final_pkg_list, key=lambda k: k['id'])
+
+        return final_pkg_list_sorted
+    
+    def env_pkgs_id(self, id):
+        # Accepts both env and workload ID, and returns pkgs for envs that match
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.env_pkgs(env_conf_id, repo_id, arch)
+        
+        # It's a workload!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.env_pkgs(env_conf_id, repo_id, arch)
+        
+        raise ValueError("That seems to be an invalid ID!")
+
+    def workload_size(self, workload_conf_id, env_conf_id, repo_id, arch):
+        # A total size of a workload (or multiple combined!)
+        pkgs = self.workload_pkgs(workload_conf_id, env_conf_id, repo_id, arch)
+        size = 0
+        for pkg in pkgs:
+            size += pkg["installsize"]
+        return size
+
+    def env_size(self, env_conf_id, repo_id, arch):
+        # A total size of an env (or multiple combined!)
+        pkgs = self.env_pkgs(env_conf_id, repo_id, arch)
+        size = 0
+        for pkg in pkgs:
+            size += pkg["installsize"]
+        return size
+
+    def workload_size_id(self, id):
+        # Accepts both env and workload ID, and returns pkgs for envs that match
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.workload_size(None, env_conf_id, repo_id, arch)
+        
+        # It's a workload!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.workload_size(workload_conf_id, env_conf_id, repo_id, arch)
+        
+        raise ValueError("That seems to be an invalid ID!")
+    
+    def env_size_id(self, id):
+        # Accepts both env and workload ID, and returns pkgs for envs that match
+        id_components = id.split(":")
+
+        # It's an env!
+        if len(id_components) == 3:
+            env_conf_id = id_components[0]
+            repo_id = id_components[1]
+            arch = id_components[2]
+            return self.env_size(env_conf_id, repo_id, arch)
+        
+        # It's a workload!
+        if len(id_components) == 4:
+            workload_conf_id = id_components[0]
+            env_conf_id = id_components[1]
+            repo_id = id_components[2]
+            arch = id_components[3]
+            return self.env_size(env_conf_id, repo_id, arch)
+        
+        raise ValueError("That seems to be an invalid ID!")
+    
+    def workload_url_slug(self, workload_conf_id, env_conf_id, repo_id, arch):
+        slug = "{workload_conf_id}--{env_conf_id}--{repo_id}--{arch}".format(
+            workload_conf_id=workload_conf_id,
+            env_conf_id=env_conf_id,
+            repo_id=repo_id,
+            arch=arch
+        )
+        return slug
+    
+    def env_url_slug(self, env_conf_id, repo_id, arch):
+        slug = "{env_conf_id}--{repo_id}--{arch}".format(
+            env_conf_id=env_conf_id,
+            repo_id=repo_id,
+            arch=arch
+        )
+        return slug
+
+    def workload_id_string(self, workload_conf_id, env_conf_id, repo_id, arch):
+        slug = "{workload_conf_id}:{env_conf_id}:{repo_id}:{arch}".format(
+            workload_conf_id=workload_conf_id,
+            env_conf_id=env_conf_id,
+            repo_id=repo_id,
+            arch=arch
+        )
+        return slug
+    
+    def env_id_string(self, env_conf_id, repo_id, arch):
+        slug = "{env_conf_id}:{repo_id}:{arch}".format(
+            env_conf_id=env_conf_id,
+            repo_id=repo_id,
+            arch=arch
+        )
+        return slug
+    
+    def url_slug_id(id):
+        return id.replace(":", "--")
+    
+
+
+
+
+###############################################################################
+### Generating html pages! ####################################################
+###############################################################################
+
+
+def _generate_html_page(template_name, template_data, page_name, settings):
+    log("Generating the '{page_name}' page...".format(
+        page_name=page_name
+    ))
+
+    output = settings["output"]
+
+    template_loader = jinja2.FileSystemLoader(searchpath="./templates/")
+    template_env = jinja2.Environment(loader=template_loader)
+
+    template = template_env.get_template("{template_name}.html".format(
+        template_name=template_name
+    ))
+
+    if template_data:
+        page = template.render(**template_data)
+    else:
+        page = template.render()
+
+    filename = ("{page_name}.html".format(
+        page_name=page_name.replace(":", "--")
+    ))
+
+    log("  Writing file...  ({filename})".format(
+        filename=filename
+    ))
+    with open(os.path.join(output, filename), "w") as file:
+        file.write(page)
+    
+    log("  Done!")
+    log("")
+
+
+def _generate_workload_pages(query):
+    log("Generating workload pages...")
+
+    # Workload overview pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for repo_id in query.workloads(workload_conf_id,None,None,None,output_change="repo_ids"):
+            template_data = {
+                "query": query,
+                "workload_conf_id": workload_conf_id,
+                "repo_id": repo_id
+            }
+
+            page_name = "workload-overview--{workload_conf_id}--{repo_id}".format(
+                workload_conf_id=workload_conf_id,
+                repo_id=repo_id
+            )
+            _generate_html_page("workload_overview", template_data, page_name, query.settings)
+    
+    # Workload detail pages
+    for workload_id in query.workloads(None,None,None,None,list_all=True):
+        workload = query.data["workloads"][workload_id]
+        
+        workload_conf_id = workload["workload_conf_id"]
+        workload_conf = query.configs["workloads"][workload_conf_id]
+
+        env_conf_id = workload["env_conf_id"]
+        env_conf = query.configs["envs"][env_conf_id]
+
+        repo_id = workload["repo_id"]
+        repo = query.configs["repos"][repo_id]
+
+
+        template_data = {
+            "query": query,
+            "workload_id": workload_id,
+            "workload": workload,
+            "workload_conf": workload_conf,
+            "env_conf": env_conf,
+            "repo": repo
+        }
+
+        page_name = "workload--{workload_id}".format(
+            workload_id=workload_id
+        )
+
+        _generate_html_page("workload", template_data, page_name, query.settings)
+    
+    # Workload compare arches pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for env_conf_id in query.workloads(workload_conf_id,None,None,None,output_change="env_conf_ids"):
+            for repo_id in query.workloads(workload_conf_id,env_conf_id,None,None,output_change="repo_ids"):
+
+                arches = query.workloads(workload_conf_id,env_conf_id,repo_id,None,output_change="arches")
+
+                workload_conf = query.configs["workloads"][workload_conf_id]
+                env_conf = query.configs["envs"][env_conf_id]
+                repo = query.configs["repos"][repo_id]
+
+                columns = {}
+                rows = set()
+                for arch in arches:
+                    columns[arch] = {}
+
+                    pkgs = query.workload_pkgs(workload_conf_id,env_conf_id,repo_id,arch)
+                    for pkg in pkgs:
+                        name = pkg["name"]
+                        rows.add(name)
+                        columns[arch][name] = pkg
+
+                template_data = {
+                    "query": query,
+                    "workload_conf_id": workload_conf_id,
+                    "workload_conf": workload_conf,
+                    "env_conf_id": env_conf_id,
+                    "env_conf": env_conf,
+                    "repo_id": repo_id,
+                    "repo": repo,
+                    "columns": columns,
+                    "rows": rows
+                }
+
+                page_name = "workload-cmp-arches--{workload_conf_id}--{env_conf_id}--{repo_id}".format(
+                    workload_conf_id=workload_conf_id,
+                    env_conf_id=env_conf_id,
+                    repo_id=repo_id
+                )
+
+                _generate_html_page("workload_cmp_arches", template_data, page_name, query.settings)
+    
+    # Workload compare envs pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for repo_id in query.workloads(workload_conf_id,None,None,None,output_change="repo_ids"):
+            for arch in query.workloads(workload_conf_id,None,repo_id,None,output_change="arches"):
+
+                env_conf_ids = query.workloads(workload_conf_id,None,repo_id,arch,output_change="env_conf_ids")
+
+                workload_conf = query.configs["workloads"][workload_conf_id]
+                repo = query.configs["repos"][repo_id]
+
+                columns = {}
+                rows = set()
+                for env_conf_id in env_conf_ids:
+                    columns[env_conf_id] = {}
+
+                    pkgs = query.workload_pkgs(workload_conf_id,env_conf_id,repo_id,arch)
+                    for pkg in pkgs:
+                        name = pkg["name"]
+                        rows.add(name)
+                        columns[env_conf_id][name] = pkg
+
+                template_data = {
+                    "query": query,
+                    "workload_conf_id": workload_conf_id,
+                    "workload_conf": workload_conf,
+                    "repo_id": repo_id,
+                    "repo": repo,
+                    "arch": arch,
+                    "columns": columns,
+                    "rows": rows
+                }
+
+                page_name = "workload-cmp-envs--{workload_conf_id}--{repo_id}--{arch}".format(
+                    workload_conf_id=workload_conf_id,
+                    repo_id=repo_id,
+                    arch=arch
+                )
+
+                _generate_html_page("workload_cmp_envs", template_data, page_name, query.settings)
+    
+    log("  Done!")
+    log("")
+
+
+def _generate_env_pages(query):
+    log("Generating env pages...")
+
+    for env_conf_id in query.envs(None,None,None,output_change="env_conf_ids"):
+        for repo_id in query.envs(env_conf_id,None,None,output_change="repo_ids"):
+            template_data = {
+                "query": query,
+                "env_conf_id": env_conf_id,
+                "repo_id": repo_id
+            }
+
+            page_name = "env-overview--{env_conf_id}--{repo_id}".format(
+                env_conf_id=env_conf_id,
+                repo_id=repo_id
+            )
+            _generate_html_page("env_overview", template_data, page_name, query.settings)
+    
+    # env detail pages
+    for env_id in query.envs(None,None,None,list_all=True):
+        env = query.data["envs"][env_id]
+
+        env_conf_id = env["env_conf_id"]
+        env_conf = query.configs["envs"][env_conf_id]
+
+        repo_id = env["repo_id"]
+        repo = query.configs["repos"][repo_id]
+
+        template_data = {
+            "query": query,
+            "env_id": env_id,
+            "env": env,
+            "env_conf": env_conf,
+            "repo": repo
+        }
+
+        page_name = "env--{env_id}".format(
+            env_id=env_id
+        )
+
+        _generate_html_page("env", template_data, page_name, query.settings)
+    
+    # env compare arches pages
+    for env_conf_id in query.envs(None,None,None,output_change="env_conf_ids"):
+        for repo_id in query.envs(env_conf_id,None,None,output_change="repo_ids"):
+
+            arches = query.envs(env_conf_id,repo_id,None,output_change="arches")
+
+            env_conf = query.configs["envs"][env_conf_id]
+            repo = query.configs["repos"][repo_id]
+
+            columns = {}
+            rows = set()
+            for arch in arches:
+                columns[arch] = {}
+
+                pkgs = query.env_pkgs(env_conf_id,repo_id,arch)
+                for pkg in pkgs:
+                    name = pkg["name"]
+                    rows.add(name)
+                    columns[arch][name] = pkg
+
+            template_data = {
+                "query": query,
+                "env_conf_id": env_conf_id,
+                "env_conf": env_conf,
+                "repo_id": repo_id,
+                "repo": repo,
+                "columns": columns,
+                "rows": rows
+            }
+
+            page_name = "env-cmp-arches--{env_conf_id}--{repo_id}".format(
+                env_conf_id=env_conf_id,
+                repo_id=repo_id
+            )
+
+            _generate_html_page("env_cmp_arches", template_data, page_name, query.settings)
+
+    log("  Done!")
+    log("")
+
+
+def _generate_config_pages(query):
+    log("Generating config pages...")
+
+    for conf_type in ["repos", "envs", "workloads", "labels", "views"]:
+        template_data = {
+            "query": query,
+            "conf_type": conf_type
+        }
+        page_name = "configs_{conf_type}".format(
+            conf_type=conf_type
+        )
+        _generate_html_page("configs", template_data, page_name, query.settings)
+
+    # Config repo pages
+    for repo_id,repo_conf in query.configs["repos"].items():
+        template_data = {
+            "query": query,
+            "repo_conf": repo_conf
+        }
+        page_name = "config-repo--{repo_id}".format(
+            repo_id=repo_id
+        )
+        _generate_html_page("config_repo", template_data, page_name, query.settings)
+    
+    # Config env pages
+    for env_conf_id,env_conf in query.configs["envs"].items():
+        template_data = {
+            "query": query,
+            "env_conf": env_conf
+        }
+        page_name = "config-env--{env_conf_id}".format(
+            env_conf_id=env_conf_id
+        )
+        _generate_html_page("config_env", template_data, page_name, query.settings)
+
+    # Config workload pages
+    for workload_conf_id,workload_conf in query.configs["workloads"].items():
+        template_data = {
+            "query": query,
+            "workload_conf": workload_conf
+        }
+        page_name = "config-workload--{workload_conf_id}".format(
+            workload_conf_id=workload_conf_id
+        )
+        _generate_html_page("config_workload", template_data, page_name, query.settings)
+
+    # Config label pages
+    for label_conf_id,label_conf in query.configs["labels"].items():
+        template_data = {
+            "query": query,
+            "label_conf": label_conf
+        }
+        page_name = "config-label--{label_conf_id}".format(
+            label_conf_id=label_conf_id
+        )
+        _generate_html_page("config_label", template_data, page_name, query.settings)
+
+    # Config view pages
+    for view_conf_id,view_conf in query.configs["views"].items():
+        template_data = {
+            "query": query,
+            "view_conf": view_conf
+        }
+        page_name = "config-view--{view_conf_id}".format(
+            view_conf_id=view_conf_id
+        )
+        _generate_html_page("config_view", template_data, page_name, query.settings)
+
+    log("  Done!")
+    log("")
+
+def _generate_repo_pages(query):
+    log("Generating repo pages...")
+
+    for repo_id, repo in query.configs["repos"].items():
+        for arch in repo["source"]["architectures"]:
+            template_data = {
+                "query": query,
+                "repo": repo,
+                "arch": arch
+            }
+            page_name = "repo--{repo_id}--{arch}".format(
+                repo_id=repo_id,
+                arch=arch
+            )
+            _generate_html_page("repo", template_data, page_name, query.settings)
+
+
+    log("  Done!")
+    log("")
+
+
+
+
+def generate_pages(query):
+    log("")
+    log("###############################################################################")
+    log("### Generating html pages! ####################################################")
+    log("###############################################################################")
+    log("")
+
+    # Copy static files
+    log("Copying static files...")
+    src_static_dir = os.path.join("templates", "_static")
+    output_static_dir = os.path.join(query.settings["output"])
+    subprocess.run(["cp", "-R", src_static_dir, output_static_dir])
+    log("  Done!")
+    log("")
+
+    # Generate the landing page
+    _generate_html_page("homepage", None, "index", query.settings)
+
+    # Generate the main menu page
+    _generate_html_page("results", None, "results", query.settings)
+
+    # Generate config pages
+    _generate_config_pages(query)
+
+    # Generate the top-level results pages
+    template_data = {
+        "query": query
+    }
+    _generate_html_page("repos", template_data, "repos", query.settings)
+    _generate_html_page("envs", template_data, "envs", query.settings)
+    _generate_html_page("workloads", template_data, "workloads", query.settings)
+    _generate_html_page("labels", template_data, "labels", query.settings)
+    _generate_html_page("views", template_data, "views", query.settings)
+
+    # Generate repo pages
+    _generate_repo_pages(query)
+
+    # Generate env_overview pages
+    _generate_env_pages(query)
+
+    # Generate workload_overview pages
+    _generate_workload_pages(query)
+
+    
+
+###############################################################################
+### Historic Data #############################################################
+###############################################################################
+
+def _save_current_historic_data(query):
+    log("Generating current historic data...")
+
+    # Where to save it
+    year = datetime.datetime.now().strftime("%Y")
+    week = datetime.datetime.now().strftime("%W")
+    filename = "historic_data-{year}-week_{week}.json".format(
+        year=year,
+        week=week
+    )
+    output_dir = os.path.join(query.settings["output"], "history")
+    file_path = os.path.join(output_dir, filename)
+
+    # What to save there
+    history_data = {}
+    history_data["date"] = str(datetime.datetime.now().strftime("%Y-%m-%d"))
+    history_data["workloads"] = {}
+    history_data["envs"] = {}
+    history_data["repos"] = {}
+
+    for workload_id in query.workloads(None,None,None,None,list_all=True):
+        workload = query.data["workloads"][workload_id]
+
+        workload_history = {}
+        workload_history["size"] = query.workload_size_id(workload_id)
+        workload_history["pkg_count"] = len(query.workload_pkgs_id(workload_id))
+
+        history_data["workloads"][workload_id] = workload_history
+    
+    for env_id in query.envs(None,None,None,list_all=True):
+        env = query.data["envs"][env_id]
+
+        env_history = {}
+        env_history["size"] = query.env_size_id(env_id)
+        env_history["pkg_count"] = len(query.env_pkgs_id(env_id))
+
+        history_data["envs"][env_id] = env_history
+
+    for repo_id in query.configs["repos"].keys():
+        history_data["repos"][repo_id] = {}
+
+        for arch, pkgs in query.data["pkgs"][repo_id].items():
+
+            repo_history = {}
+            repo_history["pkg_count"] = len(pkgs)
+            
+            history_data["repos"][repo_id][arch] = repo_history
+
+    # And save it
+    log("  Saving in: {file_path}".format(
+        file_path=file_path
+    ))
+    dump_data(file_path, history_data)
+
+    log("  Done!")
+    log("")
+
+
+def _read_historic_data(query):
+    log("Reading historic data...")
+
+    directory = os.path.join(query.settings["output"], "history")
+
+    # Do some basic validation of the filename
+    all_filenames = os.listdir(directory)
+    valid_filenames = []
+    for filename in all_filenames:
+        if bool(re.match("historic_data-....-week_...json", filename)):
+            valid_filenames.append(filename)
+    valid_filenames.sort()
+
+    # Get the data
+    historic_data = {}
+
+    for filename in valid_filenames:
+        with open(os.path.join(directory, filename), "r") as file:
+            try:
+                document = json.load(file)
+
+                date = datetime.datetime.strptime(document["date"],"%Y-%m-%d")
+                year = date.strftime("%Y")
+                week = date.strftime("%W")
+                key = "{year}-week_{week}".format(
+                    year=year,
+                    week=week
+                )
+            except (KeyError, ValueError):
+                err_log("Invalid file in historic data: {filename}. Ignoring.".format(
+                    filename=filename
+                ))
+                continue
+
+            historic_data[key] = document
+
+    return historic_data
+
+    log("  Done!")
+    log("")
+
+
+def _save_json_data_entry(entry_name, entry_data, settings):
+    log("Generating data entry for {entry_name}".format(
+        entry_name=entry_name
+    ))
+
+    output = settings["output"]
+
+    filename = ("{entry_name}.json".format(
+        entry_name=entry_name.replace(":", "--")
+    ))
+
+    log("  Writing file...  ({filename})".format(
+        filename=filename
+    ))
+
+    with open(os.path.join(output, filename), "w") as file:
+        json.dump(entry_data, file)
+    
+    log("  Done!")
+    log("")
+
+
+def _generate_chartjs_data(historic_data, query):
+
+    # Data for workload pages
+    for workload_id in query.workloads(None, None, None, None, list_all=True):
+
+        entry_data = {}
+
+        # First, get the dates as chart labels
+        entry_data["labels"] = []
+        for _,entry in historic_data.items():
+            date = entry["date"]
+            entry_data["labels"].append(date)
+
+        # Second, get the actual data for everything that's needed
+        entry_data["datasets"] = []
+
+        workload = query.data["workloads"][workload_id]
+        workload_conf_id = workload["workload_conf_id"]
+        workload_conf = query.configs["workloads"][workload_conf_id]
+
+        dataset = {}
+        dataset["data"] = []
+        dataset["label"] = workload_conf["name"]
+
+        for _,entry in historic_data.items():
+            try:
+                size = entry["workloads"][workload_id]["size"]
+
+                # The chart needs the size in MB, but just as a number
+                size_mb = "{0:.1f}".format(size/1024/1024)
+                dataset["data"].append(size_mb)
+            except KeyError:
+                dataset["data"].append("null")
+
+        entry_data["datasets"].append(dataset)
+
+        entry_name = "chartjs-data--workload--{workload_id}".format(
+            workload_id=workload_id
+        )
+        _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for workload overview pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for repo_id in query.workloads(workload_conf_id,None,None,None,output_change="repo_ids"):
+
+            entry_data = {}
+
+            # First, get the dates as chart labels
+            entry_data["labels"] = []
+            for _,entry in historic_data.items():
+                date = entry["date"]
+                entry_data["labels"].append(date)
+
+            # Second, get the actual data for everything that's needed
+            entry_data["datasets"] = []
+
+            for workload_id in query.workloads(workload_conf_id, None, repo_id, None, list_all=True):
+
+                workload = query.data["workloads"][workload_id]
+                env_conf_id = workload["env_conf_id"]
+                env_conf = query.configs["envs"][env_conf_id]
+
+                dataset = {}
+                dataset["data"] = []
+                dataset["label"] = "in {name} {arch}".format(
+                    name=env_conf["name"],
+                    arch=workload["arch"]
+                )
+
+                for _,entry in historic_data.items():
+                    try:
+                        size = entry["workloads"][workload_id]["size"]
+
+                        # The chart needs the size in MB, but just as a number
+                        size_mb = "{0:.1f}".format(size/1024/1024)
+                        dataset["data"].append(size_mb)
+                    except KeyError:
+                        dataset["data"].append("null")
+
+                entry_data["datasets"].append(dataset)
+
+            entry_name = "chartjs-data--workload-overview--{workload_conf_id}--{repo_id}".format(
+                workload_conf_id=workload_conf_id,
+                repo_id=repo_id
+            )
+            _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for workload cmp arches pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for env_conf_id in query.workloads(workload_conf_id,None,None,None,output_change="env_conf_ids"):
+            for repo_id in query.workloads(workload_conf_id,env_conf_id,None,None,output_change="repo_ids"):
+
+                workload_conf = query.configs["workloads"][workload_conf_id]
+                env_conf = query.configs["envs"][env_conf_id]
+                repo = query.configs["repos"][repo_id]
+
+                entry_data = {}
+
+                # First, get the dates as chart labels
+                entry_data["labels"] = []
+                for _,entry in historic_data.items():
+                    date = entry["date"]
+                    entry_data["labels"].append(date)
+
+                # Second, get the actual data for everything that's needed
+                entry_data["datasets"] = []
+
+                for workload_id in query.workloads(workload_conf_id,env_conf_id,repo_id,None,list_all=True):
+
+                    workload = query.data["workloads"][workload_id]
+                    env_conf_id = workload["env_conf_id"]
+                    env_conf = query.configs["envs"][env_conf_id]
+
+                    dataset = {}
+                    dataset["data"] = []
+                    dataset["label"] = "{arch}".format(
+                        arch=workload["arch"]
+                    )
+
+                    for _,entry in historic_data.items():
+                        try:
+                            size = entry["workloads"][workload_id]["size"]
+
+                            # The chart needs the size in MB, but just as a number
+                            size_mb = "{0:.1f}".format(size/1024/1024)
+                            dataset["data"].append(size_mb)
+                        except KeyError:
+                            dataset["data"].append("null")
+
+                    entry_data["datasets"].append(dataset)
+
+                entry_name = "chartjs-data--workload-cmp-arches--{workload_conf_id}--{env_conf_id}--{repo_id}".format(
+                    workload_conf_id=workload_conf_id,
+                    env_conf_id=env_conf_id,
+                    repo_id=repo_id
+                )
+                _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for workload cmp envs pages
+    for workload_conf_id in query.workloads(None,None,None,None,output_change="workload_conf_ids"):
+        for repo_id in query.workloads(workload_conf_id,None,None,None,output_change="repo_ids"):
+            for arch in query.workloads(workload_conf_id,None,repo_id,None,output_change="arches"):
+
+                workload_conf = query.configs["workloads"][workload_conf_id]
+                env_conf = query.configs["envs"][env_conf_id]
+                repo = query.configs["repos"][repo_id]
+
+                entry_data = {}
+
+                # First, get the dates as chart labels
+                entry_data["labels"] = []
+                for _,entry in historic_data.items():
+                    date = entry["date"]
+                    entry_data["labels"].append(date)
+
+                # Second, get the actual data for everything that's needed
+                entry_data["datasets"] = []
+
+                for workload_id in query.workloads(workload_conf_id,None,repo_id,arch,list_all=True):
+
+                    workload = query.data["workloads"][workload_id]
+                    repo = query.configs["repos"][repo_id]
+
+                    dataset = {}
+                    dataset["data"] = []
+                    dataset["label"] = "{repo} {arch}".format(
+                        repo=repo["name"],
+                        arch=workload["arch"]
+                    )
+
+                    for _,entry in historic_data.items():
+                        try:
+                            size = entry["workloads"][workload_id]["size"]
+
+                            # The chart needs the size in MB, but just as a number
+                            size_mb = "{0:.1f}".format(size/1024/1024)
+                            dataset["data"].append(size_mb)
+                        except KeyError:
+                            dataset["data"].append("null")
+
+                    entry_data["datasets"].append(dataset)
+
+                entry_name = "chartjs-data--workload-cmp-envs--{workload_conf_id}--{repo_id}--{arch}".format(
+                    workload_conf_id=workload_conf_id,
+                    repo_id=repo_id,
+                    arch=arch
+                )
+                _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for env pages
+    for env_id in query.envs(None, None, None, list_all=True):
+
+        entry_data = {}
+
+        # First, get the dates as chart labels
+        entry_data["labels"] = []
+        for _,entry in historic_data.items():
+            date = entry["date"]
+            entry_data["labels"].append(date)
+
+        # Second, get the actual data for everything that's needed
+        entry_data["datasets"] = []
+
+        env = query.data["envs"][env_id]
+        env_conf_id = env["env_conf_id"]
+        env_conf = query.configs["envs"][env_conf_id]
+
+        dataset = {}
+        dataset["data"] = []
+        dataset["label"] = env_conf["name"]
+
+        for _,entry in historic_data.items():
+            try:
+                size = entry["envs"][env_id]["size"]
+
+                # The chart needs the size in MB, but just as a number
+                size_mb = "{0:.1f}".format(size/1024/1024)
+                dataset["data"].append(size_mb)
+            except KeyError:
+                dataset["data"].append("null")
+
+        entry_data["datasets"].append(dataset)
+
+        entry_name = "chartjs-data--env--{env_id}".format(
+            env_id=env_id
+        )
+        _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for env overview pages
+    for env_conf_id in query.envs(None,None,None,output_change="env_conf_ids"):
+        for repo_id in query.envs(env_conf_id,None,None,output_change="repo_ids"):
+
+            entry_data = {}
+
+            # First, get the dates as chart labels
+            entry_data["labels"] = []
+            for _,entry in historic_data.items():
+                date = entry["date"]
+                entry_data["labels"].append(date)
+
+            # Second, get the actual data for everything that's needed
+            entry_data["datasets"] = []
+
+            for env_id in query.envs(env_conf_id, repo_id, None, list_all=True):
+
+                env = query.data["envs"][env_id]
+                env_conf_id = env["env_conf_id"]
+                env_conf = query.configs["envs"][env_conf_id]
+
+                dataset = {}
+                dataset["data"] = []
+                dataset["label"] = "in {name} {arch}".format(
+                    name=env_conf["name"],
+                    arch=env["arch"]
+                )
+
+                for _,entry in historic_data.items():
+                    try:
+                        size = entry["envs"][env_id]["size"]
+
+                        # The chart needs the size in MB, but just as a number
+                        size_mb = "{0:.1f}".format(size/1024/1024)
+                        dataset["data"].append(size_mb)
+                    except KeyError:
+                        dataset["data"].append("null")
+
+                entry_data["datasets"].append(dataset)
+
+            entry_name = "chartjs-data--env-overview--{env_conf_id}--{repo_id}".format(
+                env_conf_id=env_conf_id,
+                repo_id=repo_id
+            )
+            _save_json_data_entry(entry_name, entry_data, query.settings)
+    
+    # Data for env cmp arches pages
+    for env_conf_id in query.envs(None,None,None,output_change="env_conf_ids"):
+        for repo_id in query.envs(env_conf_id,None,None,output_change="repo_ids"):
+
+            env_conf = query.configs["envs"][env_conf_id]
+            env_conf = query.configs["envs"][env_conf_id]
+            repo = query.configs["repos"][repo_id]
+
+            entry_data = {}
+
+            # First, get the dates as chart labels
+            entry_data["labels"] = []
+            for _,entry in historic_data.items():
+                date = entry["date"]
+                entry_data["labels"].append(date)
+
+            # Second, get the actual data for everything that's needed
+            entry_data["datasets"] = []
+
+            for env_id in query.envs(env_conf_id,repo_id,None,list_all=True):
+
+                env = query.data["envs"][env_id]
+
+                dataset = {}
+                dataset["data"] = []
+                dataset["label"] = "{arch}".format(
+                    arch=env["arch"]
+                )
+
+                for _,entry in historic_data.items():
+                    try:
+                        size = entry["envs"][env_id]["size"]
+
+                        # The chart needs the size in MB, but just as a number
+                        size_mb = "{0:.1f}".format(size/1024/1024)
+                        dataset["data"].append(size_mb)
+                    except KeyError:
+                        dataset["data"].append("null")
+
+                entry_data["datasets"].append(dataset)
+
+            entry_name = "chartjs-data--env-cmp-arches--{env_conf_id}--{repo_id}".format(
+                env_conf_id=env_conf_id,
+                repo_id=repo_id
+            )
+            _save_json_data_entry(entry_name, entry_data, query.settings)
+
+
+def generate_historic_data(query):
+    log("")
+    log("###############################################################################")
+    log("### Historic Data #############################################################")
+    log("###############################################################################")
+    log("")
+
+    # Step 1: Save current data
+    _save_current_historic_data(query)
+
+    # Step 2: Read historic data
+    historic_data = _read_historic_data(query)
+
+    # Step 3: Generate Chart.js data
+    _generate_chartjs_data(historic_data, query)
+    
+
+
+###############################################################################
+### Main ######################################################################
+###############################################################################
 
 
 def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument("input", help="Directory with YAML configuration files. Only files ending with '.yaml' are accepted.")
-    parser.add_argument("output", help="Directory to contain the output.")
-    args = parser.parse_args()
+    settings = load_settings()
+    configs = get_configs(settings)
+    data = analyze_things(configs, settings)
 
-    configs = get_configs(args.input)
-    #dump_data("installs-cache.json", installs)
-    installs = install_and_load(configs)
-    #installs = load_data("installs-cache.json")
+    query = Query(data, configs, settings)
 
-    data = get_data(configs, installs)
+    generate_pages(query)
 
-    dump_data(os.path.join(args.output, "installs.json"), installs)
-    dump_data(os.path.join(args.output, "data.json"), data)
+    generate_historic_data(query)
 
-    date = datetime.datetime.strptime(data["timestamp"],"%d/%m/%Y %H:%M")
-    filedate = datetime.datetime.strftime(date, "%Y-%m-%d-%H%M")
-    filename = "data-{}.json".format(filedate)
-    dump_data(os.path.join(args.output, "history", filename), data)
+    #dump_data("data.json", data)
 
-    historic_data = load_historic_data(data, args.output)
-    get_historic_chart_data(historic_data, data)
 
-    generate_graphs(data, args.output)
-    generate_pages(data, args.output)
-    generate_reports_by_base(data, args.output)
-    generate_reports_by_use_case(data, args.output)
-    generate_reports_bases_releases(data, args.output)
-    generate_reports_bases_definitions(data, args.output)
-    generate_reports_use_cases_definitions(data, args.output)
 
-    
-    
+
+
+
+
+
+def tests_to_be_made_actually_useful_at_some_point_because_this_is_terribble():
+
+    settings = load_settings()
+    configs = get_configs(settings)
+    data = analyze_things(configs, settings)
+
+    query = Query(data, configs, settings)
+
+    print("")
+    print("")
+    print("")
+    print("test test test")
+    print("test test test")
+    print("test test test")
+    print("test test test")
+    print("")
+    print("")
+
+
+    # does_workload_exist(self, workload_conf_id, env_conf_id, repo_id, arch):
+
+    #   env-empty 
+    #   env-minimal 
+    #   label-eln-compose 
+    #   label-fedora-31-all 
+    #   label-fedora-rawhide-all 
+    #   repo-fedora-31 
+    #   repo-fedora-rawhide 
+    #   view-eln-compose 
+    #   workload-httpd 
+    #   workload-nginx 
+    #   x86_64
+    #   aarch64
+
+    print("Should be False:")
+    print(query.workloads("mleko","mleko","mleko","mleko"))
+    print("")
+
+    print("Should be True:")
+    print(query.workloads("workload-httpd","env-empty","repo-fedora-31","x86_64"))
+    print("")
+
+    print("Should be False:")
+    print(query.workloads("workload-httpd","env-empty","repo-fedora-31","aarch64"))
+    print("")
+
+    print("Should be True:")
+    print(query.workloads("workload-httpd","env-empty","repo-fedora-31",None))
+    print("")
+
+    print("Should be True:")
+    print(query.workloads("workload-nginx",None, None, None))
+    print("")
+
+    print("Should be True:")
+    print(query.workloads(None,"env-minimal",None,"x86_64"))
+    print("")
+
+    print("Should be True:")
+    print(query.workloads(None,"env-minimal",None,"aarch64"))
+    print("")
+
+    print("Should be False:")
+    print(query.workloads(None,"env-minimal","repo-fedora-31","x86_64"))
+    print("")
+
+    print("Should be False:")
+    print(query.workloads(None,"env-minimal","repo-fedora-31","aarch64"))
+    print("")
+
+    print("----------")
+    print("")
+    print("")
+
+    print("Should be 7:")
+    print(len(query.workloads(None,None,None,None,list_all=True)))
+    print("")
+
+    print("Should be 3:")
+    print(len(query.workloads(None,None,None,"aarch64",list_all=True)))
+    print("")
+
+    print("Should be 2:")
+    print(len(query.workloads("workload-nginx",None,None,None,list_all=True)))
+    print("")
+
+    print("----------")
+    print("")
+    print("")
+
+    print("Should be 2 workload-nginx:")
+    for id in query.workloads("workload-nginx",None,None,None,list_all=True):
+        print(id)
+    print("")
+
+    print("Should be all 7:")
+    for id in query.workloads(None,None,None,None,list_all=True):
+        print(id)
+    print("")
+
+    print("Should be all 6 rawhide:")
+    for id in query.workloads(None,None,"repo-fedora-rawhide",None,list_all=True):
+        print(id)
+    print("")
+
+    print("Should be all 2 empty rawhide:")
+    for id in query.workloads(None,"env-empty","repo-fedora-rawhide",None,list_all=True):
+        print(id)
+    print("")
+
+    print("Should be nothing:")
+    for id in query.workloads("workload-nginx","env-empty","repo-fedora-rawhide",None,list_all=True):
+        print(id)
+    print("")
+
+    print("----------")
+    print("")
+    print("")
+
+    print("Should be env-empty:repo-fedora-31:x86_64")
+    for id in query.envs_id("workload-httpd:env-empty:repo-fedora-31:x86_64", list_all=True):
+        print(id)
+    print("")
+
+    print("Should be workload-httpd:env-empty:repo-fedora-31:x86_64")
+    for id in query.workloads_id("workload-httpd:env-empty:repo-fedora-31:x86_64", list_all=True):
+        print(id)
+    print("")
+
+    print("Should be two, workload-httpd:env-minimal:repo-fedora-rawhide:x86_64 and workload-nginx:...")
+    for id in query.workloads_id("env-minimal:repo-fedora-rawhide:x86_64", list_all=True):
+        print(id)
+    print("")
+
+    print("----------")
+    print("")
+    print("")
+
+    print("Should be all 2 arches:")
+    for id in query.workloads("workload-httpd",None, None,None,list_all=True,output_change="arches"):
+        print(id)
+    print("")
+
+    print("Should be all 2 arches:")
+    for id in query.workloads("workload-nginx",None, None,None,list_all=True,output_change="arches"):
+        print(id)
+    print("")
+
+    print("Should be all 2 env_conf_ids:")
+    for id in query.workloads("workload-httpd",None, None,None,list_all=True,output_change="env_conf_ids"):
+        print(id)
+    print("")
+
+    print("Should be all 1 env_conf_id:")
+    for id in query.workloads("workload-nginx",None, None,None,list_all=True,output_change="env_conf_ids"):
+        print(id)
+    print("")
+
+    print("----------")
+    print("")
+    print("")
+
+    print("Should be 104 packages:")
+    pkgs = query.workload_pkgs("workload-nginx", "env-minimal", "repo-fedora-rawhide", "x86_64")
+    print (len(pkgs))
+    total = 0
+    env = 0
+    required = 0
+    for pkg in pkgs:
+        workload_id = "workload-nginx:env-minimal:repo-fedora-rawhide:x86_64"
+        if workload_id in pkg["q_in"]:
+            total += 1
+        if workload_id in pkg["q_required_in"]:
+            required += 1
+        if workload_id in pkg["q_env_in"]:
+            env +=1
+    print("")
+    print("Should be 104")
+    print(total)
+    print("Should be 22")
+    print(env)
+    print("Should be 1")
+    print(required)
+    print("")
+
+    print("Should be 208 packages:")
+    pkgs = query.workload_pkgs("workload-nginx", "env-minimal", "repo-fedora-rawhide", None)
+    print (len(pkgs))
+    total = 0
+    env = 0
+    required = 0
+    for pkg in pkgs:
+        workload_id = "workload-nginx:env-minimal:repo-fedora-rawhide:x86_64"
+        if workload_id in pkg["q_in"]:
+            total += 1
+        if workload_id in pkg["q_required_in"]:
+            required += 1
+        if workload_id in pkg["q_env_in"]:
+            env +=1
+    print("")
+    print("Should be 104")
+    print(total)
+    print("Should be 22")
+    print(env)
+    print("Should be 1")
+    print(required)
+    print("")
+    print("")
+    print("")
+
+
+
+
+        # q_in          - set of workload_ids including this pkg
+        # q_required_in - set of workload_ids where this pkg is required (top-level)
+        # q_env_in      - set of workload_ids where this pkg is in env
+        # size_text     - size in a human-readable format, like 6.5 MB
 
 if __name__ == "__main__":
     main()
