--- templates/report_base_releases.html
@@ -1,141 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-
-<style>
-    td,
-    th {
-        padding-left: 10px;
-        padding-right: 10px;
-    }
-
-    td {
-        white-space: nowrap !important;
-    }
-
-    .in-base th {
-        background-color: #accad6;
-        text-align: center;
-    }
-
-    .in-base td {
-        background-color: #d4e3ea;
-        color: #CC0066;
-        font-family: monospace;
-    }
-
-    tr.in-base td.all-pkgs {
-        background-color: #accad6 !important;
-    }
-
-    .not-in-base th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    .not-in-base td {
-        background-color: #fff;
-        font-family: monospace;
-    }
-
-    tr.not-in-base td.all-pkgs {
-        background-color: #e9ecef !important;
-    }
-
-    thead td,
-    thead th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    tbody tr,
-    tbody th {
-        border-bottom: #6c757d 1px solid;
-        border-top: #6c757d 1px solid;
-    }
-</style>
-
-<br>
-
-<div class="container-fluid">
-    <nav aria-label="breadcrumb">
-        <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a href="./">Home</a></li>
-            <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-            <li class="breadcrumb-item"><a href="./view--bases-definitions.html">Bases Images (Environments)</a></li>
-            <li class="breadcrumb-item"><a href="./view--bases-by-releases.html">Package Lists</a></li>
-            <li class="breadcrumb-item active" aria-current="page">{{base_name}}</li>
-        </ol>
-    </nav>
-
-
-    <h1> {{base_name}} <small class="text-muted">by releases</small> </h1>
-    <p> Generated at {{timestamp}}. </p>
-
-    <hr>
-
-    <table id="table_packages">
-        <thead>
-            <tr>
-                <td rowspan="3" style="background-color:#cad1d8">All packages in this report</td>
-                {% for image in report_data %}
-                <th colspan="2">Fedora {{image.version}}</th>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in report_data %}
-                <td colspan="2">{{image.size}}</td>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in report_data %}
-                <td colspan="2">
-                    {% if image.packages %}
-                    <a href="./graph--{{image.file_id}}.svg">graph</a>
-                    {% endif %}
-                </td>
-                {% endfor %}
-            </tr>
-            <tr class="sort_control_row">
-                <td style="background-color:#cad1d8">Name</td>
-                {% for image in report_data %}
-                <th>Name</th>
-                <th>Size</th>
-                {% endfor %}
-            </tr>
-        </thead>
-        <tbody>
-            {% for pkg in all_packages %}
-            <tr class="not-in-base">
-                <td class="all-pkgs">{{pkg}}</td>
-                {% for image in report_data %}
-                {% if pkg in image.packages %}
-                {% if pkg in image.required_pkgs %}
-                <td style="color:#CC0066; border-left: #555 1px solid">{{pkg}}</td>
-                <td style="color:#CC0066" data-order="{{image.packages[pkg].size}}">{{size_function(image.packages[pkg].size)}}</td>
-                {% else %}
-                <td style="border-left: #555 1px solid">{{pkg}}</td>
-                <td data-order="{{image.packages[pkg].size}}">{{size_function(image.packages[pkg].size)}}</td>
-                {% endif %}
-                {% else %}
-                <td style="border-left: #555 1px solid"> - </td>
-                <td data-order="0"> </td>
-                {% endif %}
-                {% endfor %}
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-
-</div>
-
-<script>
-    $(document).ready(function () {
-        $('#table_packages').DataTable({
-            paging: false
-        });
-    });
-</script>
-
-{% endblock %}
