--- templates/report_by_base.html
@@ -1,184 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-
-<style>
-    td,
-    th {
-        padding-left: 8px;
-        padding-right: 8px;
-    }
-
-    td {
-        white-space: nowrap !important;
-    }
-
-    .in-base th {
-        background-color: #b8cfb3;
-        text-align: center;
-    }
-
-    .in-base td {
-        background-color: #e3ece1;
-        font-family: monospace;
-    }
-
-    tr.in-base td.all-pkgs {
-        background-color: #b8cfb3 !important;
-    }
-
-    .not-in-base th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    .not-in-base td {
-        background-color: #fff;
-        font-family: monospace;
-    }
-
-    tr.not-in-base td.all-pkgs {
-        background-color: #e9ecef !important;
-    }
-
-    thead td,
-    thead th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    tbody tr,
-    tbody th {
-        border-bottom: #6c757d 1px solid;
-        border-top: #6c757d 1px solid;
-    }
-</style>
-
-<br>
-
-<div class="container-fluid">
-
-    <nav aria-label="breadcrumb">
-        <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a href="./">Home</a></li>
-            <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Use Cases</a></li>
-            <li class="breadcrumb-item"><a href="./view--bases-with-use-cases.html">Package Lists (by base image)</a></li>
-            <li class="breadcrumb-item active" aria-current="page">{{base.name}} (Fedora {{base.version}})</li>
-        </ol>
-    </nav>
-
-
-    <h1> {{base.name}} (Fedora {{base.version}})
-        {% if images %}
-        <small class="text-muted">
-            {% if images | length > 1 %}
-            with {{images | length}} use cases </h1>
-    {% else %}
-    with {{images[0].name}} </h1>
-    {% endif %}
-    </small>
-    {% else %}
-    </h1>
-    {% endif %}
-    <p> Generated at {{timestamp}}. </p>
-
-    <hr>
-
-    <table id="table_packages">
-        <thead>
-            <tr>
-                <td rowspan="3" style="background-color:#cad1d8"></td>
-                <td rowspan="3" style="background-color:#cad1d8">All packages in this report</td>
-                <td rowspan="3" style="background-color:#cad1d8"></td>
-                {% for image in images %}
-                <th colspan="2"><a href="./report-use-case-definition--{{image.definition_id}}--{{base.version}}.html">{{image.name}}</a></th>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in images %}
-                <td colspan="2">{{image.size}}</td>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in images %}
-                <td colspan="2">
-                    <a href="./graph--{{image.file_id}}.svg">graph</a> |
-                    <a href="./graph-simple--{{image.file_id}}.svg">simple graph</a>
-                </td>
-                {% endfor %}
-            </tr>
-            <tr class="sort_control_row">
-                <th style="background-color:#cad1d8">Flag</th>
-                <th style="background-color:#cad1d8">Name</th>
-                <th style="background-color:#cad1d8">Size</th>
-                {% for image in images %}
-                <th>Name</th>
-                <th>Why</th>
-                {% endfor %}
-            </tr>
-        </thead>
-        <tbody>
-            <!--tr class="in-base">
-                <th colspan="{{images | length + 3}}">--- Packages in base ---</th>
-            </tr-->
-            {% for pkg in base.packages %}
-            <tr class="in-base">
-                <td class="all-pkgs text-muted" data-order="0">(in base)</td>
-                <td class="all-pkgs">{{pkg}}</td>
-                <td class="all-pkgs" data-order="{{pkg_sizes_num[pkg]}}">{{pkg_sizes[pkg]}}</td>
-                {% for image in images %}
-                {% if pkg in image.pkgs_in_base %}
-                {% if pkg in image.required_pkgs %}
-                <td style="color:#CC0066">{{pkg}}</td>
-                <td style="color:#CC0066"><span class="float-right text-muted">(required)</span></td>
-                {% else %}
-                <td>{{pkg}}</td>
-                <td></td>
-                {% endif %}
-                {% else %}
-                <td> - </td>
-                <td></td>
-                {% endif %}
-                {% endfor %}
-            </tr>
-            {% endfor %}
-            <!--tr class="not-in-base">
-                <th colspan="{{images | length + 3}}">--- Packages not in base ---</th>
-            </tr-->
-            {% for pkg in extra_pkgs %}
-            <tr class="not-in-base">
-                <td class="all-pkgs" data-order="1"></td>
-                <td class="all-pkgs">{{pkg}}</td>
-                <td class="all-pkgs" data-order="{{pkg_sizes_num[pkg]}}">{{pkg_sizes[pkg]}}</td>
-                {% for image in images %}
-                {% if pkg in image.packages %}
-                {% if pkg in image.required_pkgs %}
-                <td style="color:#CC0066">{{pkg}}</td>
-                <td style="color:#CC0066"><span class="float-right text-muted">(required)</span></td>
-                {% else %}
-                <td>{{pkg}}</td>
-                <td></td>
-                {% endif %}
-                {% else %}
-                <td> - </td>
-                <td></td>
-                {% endif %}
-                {% endfor %}
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-
-</div>
-
-<script>
-    $(document).ready(function () {
-        $('#table_packages').DataTable({
-            paging: false
-        });
-    });
-</script>
-
-{% endblock %}
