--- templates/report_by_use_case.html
@@ -1,202 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-
-<style>
-    td,
-    th {
-        padding-left: 10px;
-        padding-right: 10px;
-    }
-
-    td {
-        white-space: nowrap !important;
-    }
-
-    .required th {
-        background-color: #accad6;
-        text-align: center;
-    }
-
-    .required td {
-        background-color: #d4e3ea;
-        color: #CC0066;
-        font-family: monospace;
-    }
-
-    tr.required td.all-pkgs {
-        background-color: #accad6 !important;
-    }
-
-    .not-required th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    .not-required td {
-        background-color: #fff;
-        font-family: monospace;
-    }
-
-    tr.not-required td.all-pkgs {
-        background-color: #e9ecef !important;
-    }
-
-    th.in-base  {
-        background-color: #e3ece1!important;
-    }
-
-    td.in-base  {
-        background-color: #e3ece1!important;
-    }
-
-
-    thead td,
-    thead th {
-        background-color: #e9ecef;
-        text-align: center;
-    }
-
-    tbody tr,
-    tbody th {
-        border-bottom: #6c757d 1px solid;
-        border-top: #6c757d 1px solid;
-    }
-</style>
-
-<br>
-
-<div class="container-fluid">
-    <nav aria-label="breadcrumb">
-        <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a href="./">Home</a></li>
-            <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Use Cases</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-on-bases.html">Package Lists (by use case)</a></li>
-            <li class="breadcrumb-item active" aria-current="page">{{base.name}} (Fedora {{base.version}})</li>
-        </ol>
-    </nav>
-
-
-    <h1> {{base.name}} (Fedora {{base.version}})
-        {% if images %}
-        <small class="text-muted">
-            {% if images | length > 1 %}
-            on {{images | length}} bases </h1>
-    {% else %}
-    on 1 base </h1>
-    {% endif %}
-    </small>
-    {% else %}
-    </h1>
-    {% endif %}
-    <p> Generated at {{timestamp}}. </p>
-
-    <hr>
-
-    <table id="table_packages">
-        <thead>
-            <tr>
-                <td rowspan="4" style="background-color:#cad1d8"></td>
-                <td rowspan="4" style="background-color:#cad1d8">All packages in this report</td>
-                <td rowspan="4" style="background-color:#cad1d8"></td>
-                {% for image in images %}
-                <th colspan="2">{{base.name}}</th>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in images %}
-                <th colspan="2">on <a
-                        href="./report-base-definition--{{image.definition_id}}--{{base.version}}.html">{{image.name}}</a>
-                </th>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in images %}
-                <td colspan="2">{{image.size}}</td>
-                {% endfor %}
-            </tr>
-            <tr>
-                {% for image in images %}
-                <td colspan="2">
-                    <a href="./graph--{{image.file_id}}.svg">graph</a> |
-                    <a href="./graph-simple--{{image.file_id}}.svg">simple graph</a>
-                </td>
-                {% endfor %}
-            </tr>
-            <tr class="sort_control_row">
-
-                <td style="background-color:#cad1d8">Why</td>
-                <td style="background-color:#cad1d8">Name</td>
-                <td style="background-color:#cad1d8">Size</td>
-                {% for image in images %}
-                <th>Name</th>
-                <th>Flag</th>
-                {% endfor %}
-            </tr>
-        </thead>
-        <tbody>
-            <!--tr class="required">
-                <th colspan="{{images | length + 3}}">--- Required packages ---</th>
-            </tr-->
-            {% for pkg in base.packages %}
-            <tr class="required">
-                <td class="all-pkgs text-muted" data-order="0">(required)</td>
-                <td class="all-pkgs">{{pkg}}</td>
-                <td class="all-pkgs" data-order="{{pkg_sizes_num[pkg]}}">{{pkg_sizes[pkg]}}</td>
-                {% for image in images %}
-                {% if pkg in image.required_pkgs %}
-                {% if pkg in image.pkgs_in_base %}
-                <td>{{pkg}}</td>
-                <td><span class="float-right text-muted">(in base)</span></td>
-                {% else %}
-                <td>{{pkg}}</td>
-                <td></td>
-                {% endif %}
-                {% else %}
-                <td data-order="zzzzz"> - </td>
-                <td></td>
-                {% endif %}
-                {% endfor %}
-            </tr>
-            {% endfor %}
-            <!--tr class="not-required">
-                <th colspan="{{images | length + 3}}">--- All other packages (dependencies) ---</th>
-            </tr-->
-            {% for pkg in extra_pkgs %}
-            <tr class="not-required">
-                <td class="all-pkgs" data-order="1"></td>
-                <td class="all-pkgs">{{pkg}}</td>
-                <td class="all-pkgs" data-order="{{pkg_sizes_num[pkg]}}">{{pkg_sizes[pkg]}}</td>
-                
-                {% for image in images %}
-                {% if pkg in image.packages %}
-                {% if pkg in image.pkgs_in_base %}
-                <td class="in-base">{{pkg}}</td>
-                <td class="in-base"><span class="float-right text-muted">(in base)</span></td>
-                {% else %}
-                <td>{{pkg}}</td>
-                <td></td>
-                {% endif %}
-                {% else %}
-                <td> - </td>
-                <td></td>
-                {% endif %}
-                {% endfor %}
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-
-</div>
-
-<script>
-    $(document).ready(function () {
-        $('#table_packages').DataTable({
-            paging: false
-        });
-    });
-</script>
-
-{% endblock %}
