--- templates/report_use_case_definition.html
@@ -1,161 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-
-<style>
-    div.package-name {
-        color: #CC0066;
-    }
-
-    div.install-option {
-        color: rgb(68, 68, 68);
-    }
-
-    div.package-name,
-    div.install-option {
-        font-family: monospace;
-
-        border-bottom: #6c757d 1px solid;
-
-    }
-
-    div.package-name:first-child,
-    div.install-option:first-child {
-        border-top: #6c757d 1px solid;
-    }
-</style>
-
-<br>
-
-<div class="container">
-    <nav aria-label="breadcrumb">
-        <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a href="./">Home</a></li>
-            <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Use Cases</a></li>
-            <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Details</a></li>
-            <li class="breadcrumb-item active" aria-current="page">{{use_case_definition.name}}
-            </li>
-        </ol>
-    </nav>
-
-    <h1> {{use_case_definition.name}} <small class="text-muted">details</small> </h1>
-    <p> Generated at {{data.timestamp}}. </p>
-
-    <hr>
-
-    <ul class="nav nav-tabs">
-        {% for version in use_case_definition["base_versions"] | sort %}
-        <li class="nav-item"><a class="nav-link {% if version == base_version %}active{% endif %}"
-                href="report-use-case-definition--{{use_case_definition.id}}--{{version}}.html">
-                Fedora {{version}}
-            </a></li>
-        {% endfor %}
-    </ul>
-
-    <br>
-
-    <div class="card">
-        <h5 class="card-header">
-            Size History
-        </h5>
-        <div class="card-body">
-            <canvas id="chLine"></canvas>
-        </div>
-    </div>
-
-    <br>
-
-    <div class="row">
-        <div class="col-lg-8">
-            <div class="card">
-                <h5 class="card-header">
-                    Current State
-                </h5>
-                <div class="card-body">
-                    <a href="./report-by-use-case--{{use_case_definition.id}}--{{base_version}}.html">Package lists and dependency graphs </a>
-                </div>
-            </div><br>
-
-        </div>
-        <div class="col-lg-4">
-            <div class="card">
-                <h5 class="card-header">
-                    Definition
-                    <a class="btn btn-primary btn-sm float-right"
-                        href="https://github.com/minimization/feedback-pipeline-config/blob/master/{{use_case_definition.id}}.yaml">Edit</a>
-
-                </h5>
-                <div class="card-body">
-                    <h5 class="card-title">Required Packages</h5>
-                    <div>
-                        {% for package in use_case_definition["packages"] |sort %}
-                        <div class="package-name">{{package}}</div>
-                        {% endfor %}
-                    </div>
-                </div>
-                <div class="card-body">
-                    <h5 class="card-title">Install Options</h5>
-                    <div>
-                        {% for option in use_case_definition["options"] %}
-                        <div class="install-option">{{option}}</div>
-                        {% endfor %}
-                    </div>
-                </div>
-                <div class="card-body">
-                    <h5 class="card-title">Installed on</h5>
-                    <div>
-                        {% for base in use_case_definition["base_names"][base_version] %}
-                        <div class="install-option">{{data.bases[base + ":" + base_version].name}}</div>
-                        {% endfor %}
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-
-
-</div>
-
-<script>
-    $(document).ready(function () {
-        $('#table_packages').DataTable({
-            paging: false,
-            info: false
-        });
-
-        function plot_graph(chartData) {
-            var chLine = document.getElementById("chLine");
-            if (chLine) {
-                new Chart(chLine, {
-                    type: 'line',
-                    data: chartData,
-                    options: {
-                        scales: {
-                            yAxes: [{
-                                ticks: {
-                                    beginAtZero: true,
-                                    callback: function (value, index, values) {
-                                        return value + " MB";
-                                    }
-                                }
-                            }]
-                        },
-                        legend: {
-                            display: false
-                        }
-                    }
-                });
-            }
-        }
-        
-        axios.get(`workload-historic-data--workload-httpd--env-minimal--repo-fedora-rawhide--x86_64.json`)
-            .then(response => {
-                plot_graph(response.data)
-            })
-    });
-</script>
-
-{% endblock %}
