--- templates/view_bases.html
@@ -1,31 +1,0 @@
-{% extends "view_layout.html" %}
-
-{% block active_view_bases %}active{% endblock %}
-
-
-{% block breadcrumbs %}
-<nav aria-label="breadcrumb">
-  <ol class="breadcrumb">
-    <li class="breadcrumb-item"><a href="./">Home</a></li>
-    <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-    <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-    <li class="breadcrumb-item"><a href="./view--bases-definitions.html">Base Images (Environments)</a></li>
-    <li class="breadcrumb-item active" aria-current="page">{% block breadcrumb_title %}{% endblock %}</li>
-  </ol>
-</nav>
-{% endblock %}
-
-{% block view_second_level_menu %}
-<ul class="nav nav-tabs">
-  <li class="nav-item">
-    <a class="nav-link {% block active_view_bases_definitions %}{% endblock %}"
-      href="./view--bases-definitions.html">Details</a>
-  </li>
-  <li class="nav-item">
-    <a class="nav-link {% block active_view_bases_compare_releases %}{% endblock %}"
-      href="./view--bases-by-releases.html">Package Lists</a>
-  </li>
-
-
-</ul>
-{% endblock %}
