--- templates/view_bases_compare_releases.html
@@ -1,45 +1,0 @@
-{% extends "view_bases.html" %}
-
-{% block breadcrumb_title %}Package Lists{% endblock %}
-
-{% block active_view_bases_compare_releases %}active{% endblock %}
-
-{% block view_content %}
-
-<p><i>Compare <strong>one base</strong> accross across <strong>multiple
-    Fedora releases</strong>.</i></p>
-
-<table class="table table-striped" id="table_packages">
-  <thead>
-    <tr class="sort_control_row">
-      <th>Name</th>
-      <th>Fedora Releases</th>
-    </tr>
-  </thead>
-  <tbody>
-    {% for base in data.base_definitions %}
-    <tr>
-      <td>{{base.name}}</td>
-      <td>
-        <a class="card-link" href="report-base-releases--{{base.id}}.html">All Fedora Releases</a>
-      </td>
-
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-
-
-<script>
-  $(document).ready(function () {
-    $('#table_packages').DataTable({
-      paging: false,
-      filter: false,
-      info: false
-    });
-  });
-</script>
-
-
-
-{% endblock %}
