--- templates/view_bases_definitions.html
@@ -1,45 +1,0 @@
-{% extends "view_bases.html" %}
-
-{% block breadcrumb_title %}Details{% endblock %}
-
-{% block active_view_bases_definitions %}active{% endblock %}
-
-{% block view_content %}
-
-<p><i>See the <strong>details of each base image</strong> — the required packages, install options, and more.</i></p>
-
-<table class="table table-striped" id="table_packages">
-  <thead>
-    <tr class="sort_control_row">
-      <th>Name</th>
-      <th>Fedora Releases</th>
-    </tr>
-  </thead>
-  <tbody>
-    {% for base in data.base_definitions %}
-    <tr>
-      <td>{{base.name}}</td>
-      <td>
-        {% for version in base.versions | sort %}
-        <a class="card-link" href="report-base-definition--{{base.id}}--{{version}}.html">Fedora {{version}}</a>
-        {% endfor %}
-      </td>
-
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-
-
-<script>
-  $(document).ready(function () {
-    $('#table_packages').DataTable({
-      paging: false,
-      filter: false,
-      info: false
-    });
-  });
-</script>
-
-
-{% endblock %}
