--- templates/view_use_cases.html
@@ -1,42 +1,0 @@
-{% extends "view_layout.html" %}
-
-{% block active_view_use_cases %}active{% endblock %}
-
-
-
-{% block breadcrumbs %}
-<nav aria-label="breadcrumb">
-  <ol class="breadcrumb">
-    <li class="breadcrumb-item"><a href="./">Home</a></li>
-    <li class="breadcrumb-item"><a href="./views.html">Results</a></li>
-    <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Dependency Reports</a></li>
-    <li class="breadcrumb-item"><a href="./view--use-cases-definitions.html">Use Cases</a></li>
-    <li class="breadcrumb-item active" aria-current="page">{% block breadcrumb_title %}{% endblock %}</li>
-  </ol>
-</nav>
-{% endblock %}
-
-
-{% block view_second_level_menu %}
-
-<ul class="nav nav-tabs">
-  <li class="nav-item">
-    <a class="nav-link {% block active_view_use_cases_definitions %}{% endblock %}"
-      href="./view--use-cases-definitions.html">Details</a>
-  </li>
-  <li class="nav-item">
-    <a class="nav-link {% block active_view_use_cases_compare_bases %}{% endblock %}"
-      href="./view--use-cases-on-bases.html">Package Lists (by use case)</a>
-  </li>
-  <!--li class="nav-item">
-    <a class="nav-link {% block active_view_use_cases_compare_releases %}{% endblock %}"
-      href="./view--use-cases-by-releases.html">Compare Releases</a>
-  </li-->
-  <li class="nav-item">
-    <a class="nav-link {% block active_view_use_cases_use_cases_by_base %}{% endblock %}"
-      href="./view--bases-with-use-cases.html">Package Lists (by base image)</a>
-  </li>
-
-</ul>
-
-{% endblock %}
