--- templates/view_use_cases_compare_bases.html
@@ -1,45 +1,0 @@
-{% extends "view_use_cases.html" %}
-
-{% block breadcrumb_title %}Package Lists (by use case){% endblock %}
-
-{% block active_view_use_cases_compare_bases %}active{% endblock %}
-
-
-{% block view_content %}
-
-<p><i>Compare <strong>one use case</strong> installed accross <strong>multiple base images</strong>.</i></p>
-
-<table class="table table-striped" id="table_packages">
-  <thead>
-    <tr class="sort_control_row">
-      <th>Name</th>
-      <th>Fedora Releases</th>
-    </tr>
-  </thead>
-  <tbody>
-    {% for use_case in data.use_case_definitions %}
-    <tr>
-      <td>{{use_case.name}}</td>
-      <td>
-        {% for version in use_case.base_versions |sort %}
-          <a class="card-link" href="report-by-use-case--{{use_case.id}}--{{version}}.html">Fedora {{version}}</a>
-        {% endfor %}
-      </td>
-      
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-
-
-<script>
-  $(document).ready(function () {
-    $('#table_packages').DataTable({
-      paging: false,
-      filter: false,
-      info: false
-    });
-  });
-</script>
-
-{% endblock %}
