--- templates/view_use_cases_compare_releases.html
@@ -1,17 +1,0 @@
-{% extends "view_use_cases.html" %}
-
-{% block breadcrumb_title %}Compare Releases{% endblock %}
-
-{% block active_view_use_cases_compare_releases %}active{% endblock %}
-
-
-{% block view_content %}
-
-<p><i>Compare <strong>one use case</strong> installed on <strong>a specific base image</strong> across <strong>multiple
-            Fedora releases</strong>.</i></p>
-
-<div class="alert alert-secondary" role="alert">
-    This view is to be done.
-</div>
-
-{% endblock %}
