--- templates/view_use_cases_definitions.html
@@ -1,44 +1,0 @@
-{% extends "view_use_cases.html" %}
-
-{% block breadcrumb_title %}Details{% endblock %}
-
-{% block active_view_use_cases_definitions %}active{% endblock %}
-
-{% block view_content %}
-
-<p><i>See the <strong>details of each use case</strong> — the required packages, install options, and more.</i></p>
-
-<table class="table table-striped" id="table_packages">
-    <thead>
-        <tr class="sort_control_row">
-            <th>Name</th>
-            <th>Fedora Releases</th>
-        </tr>
-    </thead>
-    <tbody>
-        {% for use_case in data.use_case_definitions %}
-        <tr>
-            <td>{{use_case.name}}</td>
-            <td>
-                <!--a href="./report-use-case-definition--{{use_case.id}}.html">All Fedora Releases</a-->
-                {% for version in use_case.base_versions | sort %}
-                <a class="card-link" href="./report-use-case-definition--{{use_case.id}}--{{version}}.html">Fedora {{version}}</a>
-                {% endfor %}
-            </td>
-        </tr>
-        {% endfor %}
-    </tbody>
-</table>
-
-
-<script>
-    $(document).ready(function () {
-        $('#table_packages').DataTable({
-            paging: false,
-            filter: false,
-            info: false
-        });
-    });
-</script>
-
-{% endblock %}
