--- templates/view_use_cases_use_cases_by_base.html
@@ -1,45 +1,0 @@
-{% extends "view_use_cases.html" %}
-
-{% block breadcrumb_title %}Package Lists (by base image){% endblock %}
-
-{% block active_view_use_cases_use_cases_by_base %}active{% endblock %}
-
-{% block view_content %}
-
-<p><i>Compare <strong>multiple use cases</strong> installed on a <strong>specific base image</strong>.</i></p>
-
-<table class="table table-striped" id="table_packages">
-  <thead>
-    <tr class="sort_control_row">
-      <th>Name</th>
-      <th>Fedora Releases</th>
-    </tr>
-  </thead>
-  <tbody>
-    {% for base in data.base_definitions %}
-    <tr>
-      <td>All use cases on the {{base.name}}</td>
-      <td>
-          {% for version in base.versions | sort %}
-        <a class="card-link" href="report-by-base--{{base.id}}--{{version}}.html">Fedora {{version}}</a>
-        {% endfor %}
-      </td>
-
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-
-
-<script>
-  $(document).ready(function () {
-    $('#table_packages').DataTable({
-      paging: false,
-      filter: false,
-      info: false
-    });
-  });
-</script>
-
-
-{% endblock %}
